Use bootstrapping method to calculate the variance associated to the GSVA score obtained for each sample

Use bootstrapping method to calculate the variance associated to the GSVA score obtained for each sample

bootstrapGSVA(geneLists, countMatrix, bootstrapRatio, bootstrapNbr)

Arguments

geneLists

a list containing signature gene lists.

countMatrix

a normalized gene expression data matrix with rows corresponding to genes and columns to samples. The matrix has been filtered to contained only the genes present in the signature gene lists.

bootstrapRatio

a numeric between 0 and 1 representing the number of samples that are retained for the bootstrap step.

Value

a list containing the variance calculated from the bootstrapping method on each sample for each signature gene list. The list contains one entry per signature gene list.

Author

Astrid Deschênes

Examples

## Load signatures
data(signatures)

## Create a list of two signatures
geneLists <- list()
geneLists[["classical"]] <- 
    signatures$`2018_Tiriac_PDAC_PDO_classical_signature`
geneLists[["basal"]] <- 
    signatures$`2018_Tiriac_PDAC_PDO_basal-like_signature`

## Load demo normalized expected counts
data(expNormalCountsDemo)

## Calculate variance for each sample using boostrap method
splitTypeR:::bootstrapGSVA(geneLists=geneLists, 
    countMatrix=expNormalCountsDemo, bootstrapRatio=0.75, 
    bootstrapNbr=20)
#> $classical
#>                   [,1]        [,2]        [,3]        [,4]       [,5]
#> Patient_1           NA  0.19554656  0.13779321  0.35170089  0.1530190
#> Patient_2   0.46595672  0.45301040  0.40093918  0.50746354  0.4155694
#> Patient_3           NA -0.72588832 -0.74677003 -0.70140612 -0.7434033
#> Patient_4   0.44840827  0.35515741  0.32307692          NA  0.3254485
#> Patient_5  -0.67799490          NA -0.67414988 -0.61122125 -0.6317726
#> Patient_6   0.58287074  0.59080095  0.56129149  0.63713820  0.5372247
#> Patient_7  -0.08533453 -0.10091265 -0.18064873  0.08266549 -0.1975854
#> Patient_8           NA -0.05879619 -0.09960705  0.06380663 -0.2186847
#> Patient_9   0.03048327 -0.05302549          NA  0.05736049 -0.0686772
#> Patient_10          NA  0.46462470  0.37591285  0.50437197         NA
#> Patient_11  0.62154624  0.56658395  0.58447576          NA  0.5644541
#> Patient_12 -0.17220959          NA -0.22640761 -0.05571267 -0.1547992
#> Patient_13  0.32384615  0.29680857          NA          NA  0.2334349
#> Patient_14  0.37170437          NA  0.31355253          NA  0.3098752
#> Patient_15  0.11383583          NA  0.01430976  0.21107099         NA
#> Patient_16 -0.35905289 -0.32673609 -0.38277113 -0.14848415 -0.3111172
#> Patient_17 -0.66148242 -0.68222348          NA -0.66106433 -0.7026266
#> Patient_18          NA          NA  0.70465773          NA  0.7315271
#> Patient_19 -0.01214575          NA -0.07851038  0.15692308 -0.1327370
#> Patient_20 -0.46299787 -0.44787692 -0.50351346          NA         NA
#> Patient_21          NA  0.54516995          NA  0.64962280         NA
#> Patient_22          NA -0.26636407          NA -0.12537538         NA
#> Patient_23  0.53815061          NA  0.50907555          NA  0.5104344
#> Patient_24  0.43484722          NA          NA  0.60072152  0.4333293
#> Patient_25 -0.90424453 -0.88596941 -0.92307692 -0.86476661         NA
#> Patient_26 -0.46684018 -0.46923741 -0.56449106 -0.39742533 -0.5652302
#> Patient_27  0.77814570  0.75334448  0.74755700          NA         NA
#> Patient_28 -0.17020448 -0.21286315          NA -0.20673077 -0.2882705
#> Patient_29          NA  0.70416141  0.63383390  0.74502625  0.6899335
#> Patient_30  0.41941972  0.42129492          NA  0.47301952         NA
#>                   [,6]        [,7]         [,8]         [,9]       [,10]
#> Patient_1   0.04746108  0.19485222           NA           NA  0.08395604
#> Patient_2   0.39821175  0.35895568  0.386567164           NA  0.27581627
#> Patient_3  -0.75171821 -0.75063830           NA -0.707357860          NA
#> Patient_4   0.28890662  0.35510980  0.426546003           NA  0.25834543
#> Patient_5           NA -0.70633562 -0.653615845 -0.643552709          NA
#> Patient_6   0.50627943  0.55607615  0.587648801           NA  0.42651045
#> Patient_7  -0.17357341 -0.05061605 -0.112509583  0.052511125 -0.27349004
#> Patient_8           NA -0.07365472 -0.114422348  0.066599811          NA
#> Patient_9  -0.18293960          NA           NA  0.007029375 -0.23758626
#> Patient_10          NA  0.41614611  0.412012645  0.464506265  0.37727246
#> Patient_11  0.54718283  0.58701805           NA  0.662869576  0.49497235
#> Patient_12 -0.22687853 -0.15874233 -0.146284830 -0.106626047 -0.31549262
#> Patient_13  0.12228117          NA  0.235297533  0.466962895  0.12672248
#> Patient_14  0.27455717          NA  0.349858442  0.462477980          NA
#> Patient_15          NA  0.02313574  0.101461601  0.188077907 -0.07413882
#> Patient_16          NA          NA -0.393247211 -0.172205438 -0.48142632
#> Patient_17 -0.67202414 -0.69953251 -0.664744179 -0.658662572          NA
#> Patient_18  0.67252112  0.66851639  0.686836449           NA  0.66851898
#> Patient_19          NA          NA -0.005078705  0.126175549 -0.23082672
#> Patient_20 -0.48017317 -0.49667692 -0.448768623           NA          NA
#> Patient_21  0.42542771  0.44869542           NA  0.571239757  0.34893810
#> Patient_22 -0.32680363          NA -0.235909117 -0.095659258 -0.37087275
#> Patient_23  0.46325180  0.50409035  0.533096263  0.581436901  0.42835945
#> Patient_24          NA  0.44366714  0.447274776           NA          NA
#> Patient_25 -0.92307692 -0.92307692 -0.884615385 -0.846261783 -0.88461538
#> Patient_26 -0.53569072 -0.49076657           NA -0.412846090 -0.59595559
#> Patient_27  0.73611111  0.73431133           NA           NA  0.70910556
#> Patient_28          NA          NA           NA -0.076200942 -0.38894269
#> Patient_29  0.64011872  0.66443701  0.701130856  0.750825083          NA
#> Patient_30  0.39239143          NA  0.393116157  0.438180262  0.36354541
#>                  [,11]       [,12]       [,13]       [,14]       [,15]
#> Patient_1   0.16034806  0.20140545  0.33611769          NA  0.20077049
#> Patient_2   0.42030908  0.41885965  0.45189454  0.46073197          NA
#> Patient_3  -0.77327226 -0.72694942 -0.73922232          NA          NA
#> Patient_4           NA  0.37234799  0.40790120          NA          NA
#> Patient_5  -0.66758965 -0.64382126 -0.65852213 -0.67146213 -0.69372221
#> Patient_6   0.49914052          NA          NA  0.58397385  0.55648943
#> Patient_7           NA  0.00616023 -0.06823821 -0.11634455 -0.19457014
#> Patient_8  -0.18825911 -0.13737097          NA -0.09862909          NA
#> Patient_9           NA -0.11150472 -0.14230769          NA -0.18004295
#> Patient_10  0.43057861  0.40195909  0.43222308  0.38455746  0.37979654
#> Patient_11          NA  0.62201915  0.59392713  0.59749479  0.56487640
#> Patient_12          NA          NA -0.11840232 -0.17086029 -0.23338485
#> Patient_13  0.28111078  0.31115731  0.33133882  0.27148641          NA
#> Patient_14  0.31662689          NA  0.33899959  0.33325314  0.26983267
#> Patient_15  0.10655738  0.03182351          NA  0.11960090  0.01810328
#> Patient_16 -0.42773074 -0.30522756 -0.35961965 -0.34005305 -0.40898705
#> Patient_17 -0.66241717          NA -0.71399495          NA -0.70749361
#> Patient_18  0.64529559          NA  0.73562484          NA  0.71338727
#> Patient_19 -0.06108776 -0.05957768          NA  0.02140009          NA
#> Patient_20 -0.47255508 -0.50402930          NA -0.50231623          NA
#> Patient_21  0.46328671  0.48081825  0.54038944  0.45163420  0.43728056
#> Patient_22 -0.19828922          NA -0.21475086 -0.19991741 -0.25620203
#> Patient_23  0.50587523  0.54390304          NA  0.53070148  0.50310370
#> Patient_24          NA  0.44559245  0.46077394          NA          NA
#> Patient_25          NA -0.88461538 -0.88461538 -0.86508050 -0.88461538
#> Patient_26 -0.54711287          NA -0.48625452 -0.50630413 -0.53491685
#> Patient_27  0.76320132          NA          NA          NA  0.72896440
#> Patient_28 -0.25544028 -0.24718428          NA -0.26308695 -0.36640802
#> Patient_29  0.66809353  0.68592933  0.68938576  0.69611638  0.64596849
#> Patient_30          NA  0.39698015  0.42541702  0.41398601  0.38697017
#>                 [,16]        [,17]       [,18]       [,19]       [,20]
#> Patient_1          NA  0.259852997          NA          NA  0.18240480
#> Patient_2   0.3042010  0.505051665  0.36123566          NA          NA
#> Patient_3  -0.7827225 -0.717283142 -0.76455026          NA -0.75152089
#> Patient_4          NA  0.347746203  0.14825456          NA  0.40098857
#> Patient_5          NA -0.617786377          NA -0.69150780 -0.67723265
#> Patient_6   0.4649303  0.598604679  0.40778499  0.56813621          NA
#> Patient_7  -0.1999595           NA -0.28382433 -0.16115023          NA
#> Patient_8          NA -0.058188397          NA -0.13692308 -0.13834072
#> Patient_9  -0.1465341 -0.071032061 -0.22958762          NA -0.16935576
#> Patient_10  0.3885921  0.441061254  0.28398058  0.42423522  0.38552752
#> Patient_11         NA  0.595780433  0.48126314  0.56011207  0.59353147
#> Patient_12 -0.2990214 -0.166627310          NA -0.22809787 -0.28422121
#> Patient_13  0.1713233  0.302484286  0.08393734  0.20471079  0.22487279
#> Patient_14  0.2239894  0.354109343  0.17419701  0.28642354  0.33434761
#> Patient_15         NA           NA -0.03291934  0.00466007  0.07951707
#> Patient_16         NA -0.360636277          NA -0.41155022 -0.43710270
#> Patient_17 -0.7216561 -0.682013154 -0.70442954 -0.67049053          NA
#> Patient_18  0.6348564  0.757425743  0.59584961  0.68544419  0.71194763
#> Patient_19 -0.1309354 -0.006884166 -0.28963159 -0.13814757          NA
#> Patient_20 -0.5448342           NA -0.58799437 -0.50594322          NA
#> Patient_21  0.4261900           NA  0.41189185          NA  0.46697681
#> Patient_22 -0.2955730           NA -0.35049834 -0.28979907 -0.30259963
#> Patient_23  0.4485225  0.535666645  0.31637717  0.49073105  0.51469317
#> Patient_24  0.3600302           NA          NA  0.35169065  0.34814728
#> Patient_25         NA -0.884615385          NA -0.88656007 -0.88461538
#> Patient_26 -0.5843565 -0.502393737 -0.61336032          NA -0.53221030
#> Patient_27  0.7256662           NA  0.72439024  0.76973684  0.76418152
#> Patient_28 -0.3362495 -0.188869078          NA          NA          NA
#> Patient_29  0.6439308  0.716456494  0.62510121  0.63440000          NA
#> Patient_30  0.3067744           NA  0.27512902  0.38622043  0.37074500
#> 
#> $basal
#>                   [,1]        [,2]        [,3]         [,4]       [,5]
#> Patient_1           NA -0.19093377 -0.22888889 -0.348148148 -0.1760417
#> Patient_2  -0.43693143 -0.45120968 -0.43449612 -0.494047619 -0.4213411
#> Patient_3           NA  0.80634683  0.79815891  0.779687500  0.7955882
#> Patient_4  -0.54124108 -0.47048611 -0.39865320           NA -0.3930365
#> Patient_5   0.83035714          NA  0.82923977  0.757177616  0.7927039
#> Patient_6  -0.66088435 -0.62721088 -0.60874126 -0.702083333 -0.6132985
#> Patient_7   0.06013825  0.11860987  0.14141717 -0.069841270  0.2001464
#> Patient_8           NA  0.19476190  0.17309994 -0.007773386  0.2482143
#> Patient_9   0.01190476  0.04478764          NA -0.038759690  0.1331715
#> Patient_10          NA -0.47234848 -0.42094017 -0.558470445         NA
#> Patient_11 -0.68733032 -0.59427083 -0.63008896           NA -0.6689655
#> Patient_12  0.07436353          NA  0.19251701 -0.028936605  0.1109253
#> Patient_13 -0.40391621 -0.23414634          NA           NA -0.2412935
#> Patient_14 -0.43193961          NA -0.32884828           NA -0.3884058
#> Patient_15 -0.09939117          NA -0.05151286 -0.185264228         NA
#> Patient_16  0.38885820  0.39453125  0.47985612  0.345402299  0.4655172
#> Patient_17  0.76560142  0.74173228          NA  0.712073491  0.7829502
#> Patient_18          NA          NA -0.82435897           NA -0.8596672
#> Patient_19  0.07950329          NA  0.13797397 -0.161566485  0.2483900
#> Patient_20  0.55616438  0.62634947  0.55816993           NA         NA
#> Patient_21          NA -0.65813953          NA -0.762703252         NA
#> Patient_22          NA  0.25632799          NA  0.181044487         NA
#> Patient_23 -0.62973856          NA -0.49946237           NA -0.6016064
#> Patient_24 -0.48738739          NA          NA -0.604142948 -0.3610155
#> Patient_25  0.94267260  0.90551378  0.96305732  0.895301306         NA
#> Patient_26  0.51012270  0.61449275  0.62460918  0.464150943  0.6143068
#> Patient_27 -0.79482579 -0.81559469 -0.79315620           NA         NA
#> Patient_28  0.33644578  0.26496018          NA  0.173183583  0.3565263
#> Patient_29          NA -0.75071542 -0.68282828 -0.821349446 -0.7290221
#> Patient_30 -0.47194776 -0.40306859          NA -0.499305556         NA
#>                  [,6]        [,7]        [,8]         [,9]      [,10]
#> Patient_1  -0.1235988 -0.26716418          NA           NA -0.1185897
#> Patient_2  -0.3870445 -0.42983193 -0.46791094           NA -0.3330065
#> Patient_3   0.8047445  0.79945598          NA  0.765952016         NA
#> Patient_4  -0.3884787 -0.38305322 -0.49065041           NA -0.2632871
#> Patient_5          NA  0.82657856  0.78186275  0.789583333         NA
#> Patient_6  -0.5940580 -0.59746377 -0.66319444           NA -0.4995338
#> Patient_7   0.1687773  0.10667627  0.09379845 -0.129577465  0.2848765
#> Patient_8          NA  0.22344633  0.14968553  0.035015291         NA
#> Patient_9   0.2459203          NA          NA -0.050629556  0.3317992
#> Patient_10         NA -0.45167866 -0.47305296 -0.564487633 -0.3860606
#> Patient_11 -0.6339219 -0.65851970          NA -0.725476558 -0.6410667
#> Patient_12  0.2868231  0.22981651  0.21394375 -0.003482587  0.3387125
#> Patient_13 -0.1411765          NA -0.25725191 -0.440310078 -0.1190748
#> Patient_14 -0.3155409          NA -0.39941725 -0.469882498         NA
#> Patient_15         NA -0.02886905 -0.08588110 -0.155917160  0.1164145
#> Patient_16         NA          NA  0.45862709  0.342380952  0.5983709
#> Patient_17  0.7509296  0.75884666  0.74055118  0.749062918         NA
#> Patient_18 -0.7694880 -0.76912908 -0.80868263           NA -0.7404646
#> Patient_19         NA          NA  0.09332130 -0.062565445  0.3555773
#> Patient_20  0.6297218  0.60297718  0.55981735           NA         NA
#> Patient_21 -0.5442126 -0.58670886          NA -0.709002433 -0.5292486
#> Patient_22  0.3867201          NA  0.28922306  0.064579380  0.4256595
#> Patient_23 -0.4804997 -0.53421053 -0.63160173 -0.688638800 -0.4774864
#> Patient_24         NA -0.45536723 -0.43333333           NA         NA
#> Patient_25  0.9569620  0.95696203  0.94723295  0.916666667  0.9558442
#> Patient_26  0.6776720  0.62563917          NA  0.509184727  0.7233577
#> Patient_27 -0.8428000 -0.78102894          NA           NA -0.7698465
#> Patient_28         NA          NA          NA  0.206089744  0.4767589
#> Patient_29 -0.6940771 -0.73904448 -0.74108020 -0.812689691         NA
#> Patient_30 -0.3018354          NA -0.42342342 -0.498138748 -0.3797048
#>                  [,11]       [,12]       [,13]       [,14]        [,15]
#> Patient_1  -0.25303030 -0.26744248 -0.28292683          NA -0.133016376
#> Patient_2  -0.39498382 -0.42373527 -0.51117764 -0.40565093           NA
#> Patient_3   0.83696370  0.82408223  0.80015015          NA           NA
#> Patient_4           NA -0.38932244 -0.48174204          NA           NA
#> Patient_5   0.80821596  0.77592593  0.79082232  0.81492537  0.819684111
#> Patient_6  -0.53635347          NA          NA -0.62998847 -0.579977247
#> Patient_7           NA  0.07474120  0.04322917  0.08405461  0.100278199
#> Patient_8   0.25401737  0.22610063          NA  0.17711443           NA
#> Patient_9           NA  0.09490644  0.13394683          NA  0.207923497
#> Patient_10 -0.52519509 -0.44481193 -0.51344086 -0.43390915 -0.488968825
#> Patient_11          NA -0.66935484 -0.68876221 -0.69206349 -0.642937400
#> Patient_12          NA          NA  0.16285141  0.17212389  0.295045045
#> Patient_13 -0.25433213 -0.24413282 -0.26054688 -0.30924855           NA
#> Patient_14 -0.38355856          NA -0.42469586 -0.38947368 -0.336573263
#> Patient_15 -0.05869565 -0.05757121          NA -0.06253823  0.006008359
#> Patient_16  0.51220657  0.44585799  0.37369478  0.46029412  0.486815920
#> Patient_17  0.74513444          NA  0.74296149          NA  0.738888889
#> Patient_18 -0.79259259          NA -0.85620155          NA -0.784388186
#> Patient_19  0.23953356  0.07989165          NA  0.14702048           NA
#> Patient_20  0.62264631  0.61029994          NA  0.56394094           NA
#> Patient_21 -0.63829787 -0.65744048 -0.63967280 -0.58274854 -0.623498965
#> Patient_22  0.23706140          NA  0.25982256  0.31896024  0.303717026
#> Patient_23 -0.45561313 -0.62823691          NA -0.62000000 -0.494871795
#> Patient_24          NA -0.42887701 -0.42473814          NA           NA
#> Patient_25          NA  0.94238156  0.92462312  0.93294574  0.949677419
#> Patient_26  0.72792023          NA  0.56659056  0.56458658  0.634108527
#> Patient_27 -0.81425955          NA          NA          NA -0.750842849
#> Patient_28  0.38254344  0.32566372          NA  0.32792069  0.436158192
#> Patient_29 -0.73875969 -0.77335407 -0.76948798 -0.78146948 -0.730684327
#> Patient_30          NA -0.48198198 -0.45555556 -0.44452055 -0.365797101
#>                 [,16]       [,17]       [,18]       [,19]        [,20]
#> Patient_1          NA -0.30170604          NA          NA -0.197674419
#> Patient_2  -0.3950495 -0.45111329 -0.40606061          NA           NA
#> Patient_3   0.8367947  0.78575668  0.84976690          NA  0.807594937
#> Patient_4          NA -0.46343451 -0.19580574          NA -0.384988962
#> Patient_5          NA  0.75316092          NA  0.82458034  0.807271010
#> Patient_6  -0.5028497 -0.63245614 -0.48079470 -0.63207763           NA
#> Patient_7   0.2423729          NA  0.32582062  0.15147059           NA
#> Patient_8          NA  0.15414313          NA  0.16138434  0.247395833
#> Patient_9   0.2705993  0.10926680  0.30100548          NA  0.154040404
#> Patient_10 -0.4934708 -0.47371911 -0.46122893 -0.46732852 -0.447260274
#> Patient_11         NA -0.68395867 -0.51869159 -0.67732709 -0.663857374
#> Patient_12  0.3766886  0.06849817          NA  0.23539519  0.223083779
#> Patient_13 -0.1654412 -0.28936170 -0.08983740 -0.15807990 -0.302482270
#> Patient_14 -0.3178478 -0.39449378 -0.26666667 -0.33821549 -0.337179487
#> Patient_15         NA          NA  0.07625538 -0.02556936  0.002160494
#> Patient_16         NA  0.45940325          NA  0.50616143  0.465408805
#> Patient_17  0.7851351  0.73639500  0.80089127  0.74523810           NA
#> Patient_18 -0.7843558 -0.85133438 -0.76666667 -0.77143217 -0.787320574
#> Patient_19  0.3228451  0.03900709  0.40590829  0.19188406           NA
#> Patient_20  0.6954128          NA  0.68650180  0.60000000           NA
#> Patient_21 -0.6104762          NA -0.50521978          NA -0.583703704
#> Patient_22  0.2583333          NA  0.34302326  0.34684685  0.347272727
#> Patient_23 -0.4314711 -0.61118077 -0.33798450 -0.53833333 -0.554726368
#> Patient_24 -0.3592334          NA          NA -0.42757576 -0.323556797
#> Patient_25         NA  0.94480103          NA  0.94238156  0.938775510
#> Patient_26  0.6963768  0.52424242  0.80049365          NA  0.627450980
#> Patient_27 -0.8154989          NA -0.81650512 -0.79573460 -0.785365854
#> Patient_28  0.4624595  0.31120645          NA          NA           NA
#> Patient_29 -0.7113576 -0.76420244 -0.66596091 -0.70973845           NA
#> Patient_30 -0.2718468          NA -0.21499162 -0.39594595 -0.413471503
#>