Assign the typing for each sample and each gene signature

This function assigns the typing for all samples in each signature using the two normal distributions obtained for each signature. One normal distribution is associated with the samples with the signatures while the other normal distribution is called the alternative distribution. The null hypothesis is that the sample is part of the alternative distribution. At this step, the sample is considered "unclassified". The null hypothesis is rejected when the probability of the sample being part of the alternative distribution is inferior to 0.05. The sample is then considered to be part of the distribution associated to the signature and is assigned to the signature.

classification(geneLists, gsvaRes, modelMix)

Arguments

geneLists

a list containing signature gene lists.

gsvaRes

a matrix as generated by the GSVA function. The matrix should contain the GSVA score for all the samples (column) and the signatures (row).

modelMix

a TODO

Value

a list TODO

Author

Astrid Deschênes

Examples

## Required library
library(GSVA)

## Load signatures
data(signatures)

## Create a list of two signatures
geneLists <- list()
geneLists[["classical"]] <- 
    signatures$`2018_Tiriac_PDAC_PDO_classical_signature`
geneLists[["basal"]] <- 
    signatures$`2018_Tiriac_PDAC_PDO_basal-like_signature`

## Load demo normalized expected counts
data(expNormalCountsDemo)

gsvaParameter <- gsvaParam(exprData=as.matrix(expNormalCountsDemo),
    geneSets=geneLists)
resClass <- gsva(param=gsvaParameter, verbose=FALSE)

## Calculate standard deviation for each sample using the permutation method
permResult <- splitTypeR:::permuteGSVA(geneLists=geneLists, 
    countMatrix=expNormalCountsDemo, permRatio=0.75, permNbr=20)

## Draw 10 values from the normal distribution for each sample (upscaling)
normResult <- splitTypeR:::extractFromNormalDist(geneLists=geneLists, 
    gsvaRes=resClass, gsvaSD=permResult$SD, nbValues=10)

## Obtain the mixtures of normal distributions for each signature
model <- splitTypeR:::normalMix(geneLists=geneLists, resultNorm=normResult)
#> number of iterations= 35 
#> number of iterations= 28 

## Obtain the assignment of each sample for each signature
result <- splitTypeR:::classification(geneLists=geneLists, gsvaRes=resClass, 
    modelMix=model)
## Assignment for the classical signature
head(result$classical)
#>           GSVA_score assignedType
#> Patient_1  0.1422663 unclassified
#> Patient_2  0.4274225    classical
#> Patient_3 -0.7532768 unclassified
#> Patient_4  0.3479971    classical
#> Patient_5 -0.6732937 unclassified
#> Patient_6  0.5735090    classical
## Assignment for the basal signature
head(result$basal)
#>           GSVA_score assignedType
#> Patient_1 -0.2385417 unclassified
#> Patient_2 -0.4403292 unclassified
#> Patient_3  0.8006550        basal
#> Patient_4 -0.4500838 unclassified
#> Patient_5  0.8175287        basal
#> Patient_6 -0.6076389 unclassified