Sample a specific number of values from the normal distribution associated with each sample and each gene signature

This function generates a normal distribution for each sample and gene signature, using the actual GSVA score for the mean and the permutation results for the standard deviation. The function returns a user-defined number of values sampled from this distribution.

extractFromNormalDist(geneLists, gsvaRes, gsvaSD, nbValues)

Arguments

geneLists

a list containing signature gene lists.

gsvaRes

a matrix as generated by the GSVA function. The matrix should contain the GSVA score for all the samples (column) and the signatures (row).

gsvaSD

a list containing one entry per gene signature. Each entry contains a vector of the standard deviations for each sample as calculated from the permutations.

nbValues

a integer superior or egal to 2 representing the number of values sampled.

Value

a list containing one entry for each gene signature present in the geneLists object. Each entry contains a matrix with all the values sampled from the normal distribution obtained from the GSVA score (mean) and the permutations (standard deviation). The rows correspond to the samples, while the columns correspond to the number of draws.

Author

Astrid Deschênes

Examples

## Required library
library(GSVA)

## Load signatures
data(signatures)

## Create a list of two signatures
geneLists <- list()
geneLists[["classical"]] <- 
    signatures$`2018_Tiriac_PDAC_PDO_classical_signature`
geneLists[["basal"]] <- 
    signatures$`2018_Tiriac_PDAC_PDO_basal-like_signature`

## Load demo normalized expected counts
data(expNormalCountsDemo)

## Fix seed for reproducibility
set.seed(144)

## Calculate the real GSVA values
gsvaParameter <- gsvaParam(exprData=as.matrix(expNormalCountsDemo), 
     geneSets=geneLists)
resGSVA <- GSVA::gsva(param=gsvaParameter, verbose=FALSE)

## Calculate standard deviation for each sample using the permutation method
permResult <- splitTypeR:::permuteGSVA(geneLists=geneLists, 
    countMatrix=expNormalCountsDemo, permRatio=0.75, permNbr=20)

## For each signature, randomly extract values from the normal 
## distribution for each sample
normResult <- splitTypeR:::extractFromNormalDist(geneLists=geneLists, 
    gsvaRes=resGSVA, gsvaSD=permResult$SD, nbValues=10)
## For the classical signature, the sampled values for the first 3 patients
head(normResult$classical[1:3, ])
#>                  [,1]        [,2]       [,3]       [,4]        [,5]       [,6]
#> Patient_1  0.06552621 -0.04744088  0.2579767  0.1401724  0.04282027  0.0736576
#> Patient_2  0.50276388  0.48996602  0.4534817  0.2593450  0.59265803  0.3882175
#> Patient_3 -0.74669360 -0.73819720 -0.7839284 -0.7288436 -0.73232883 -0.7401041
#>                 [,7]       [,8]       [,9]      [,10]
#> Patient_1  0.1773961  0.2352225  0.2480705  0.2423369
#> Patient_2  0.4912494  0.4785465  0.4901619  0.5812916
#> Patient_3 -0.7637563 -0.7414708 -0.7394799 -0.7666406