TODO

This function calculates TODO

normalMix(geneLists, resultNorm)

Arguments

geneLists

a list containing signature gene lists.

resultNorm

a TODO

Value

a list TODO

Author

Astrid Deschênes

Examples

## Required library
library(GSVA)

## Load signatures
data(signatures)

## Create a list of two signatures
geneLists <- list()
geneLists[["classical"]] <- 
    signatures$`2018_Tiriac_PDAC_PDO_classical_signature`
geneLists[["basal"]] <- 
    signatures$`2018_Tiriac_PDAC_PDO_basal-like_signature`

## Load demo normalized expected counts
data(expNormalCountsDemo)

## Fix seed for reproducibility
set.seed(144)

## Calculate GSVA scores for the signatures
gsvaParameter <- gsvaParam(exprData=as.matrix(expNormalCountsDemo),
    geneSets=geneLists)
resClass <- gsva(param=gsvaParameter, verbose=FALSE)

## Calculate variance for each sample using the permutation method
permResult <- splitTypeR:::permuteGSVA(geneLists=geneLists, 
    countMatrix=expNormalCountsDemo, permRatio=0.75, permNbr=20)

## For each sample, randomly select 10 values (nbValues) from the normal 
## curve where mean is the real GSVA value, and the standard deviation is 
## calculated from the permutation step.
normResult <- splitTypeR:::extractFromNormalDist(geneLists=geneLists, 
    gsvaRes=resClass, gsvaSD=permResult$SD, nbValues=10)

## Extracting the mixtures of two normal distributions for each signature
model <- splitTypeR:::normalMix(geneLists=geneLists, resultNorm=normResult)
#> number of iterations= 47 
#> number of iterations= 55 

## The two means of the normal distributions for the basal signature
model[["basal"]]$mu
#> [1] -0.5352405  0.3578356