TODO

TODO.

runSubtyping(
  geneLists,
  expectedCountsMatrix,
  permRatio = 0.75,
  permNbr = 20,
  upscaleNbr = 10
)

Arguments

geneLists

a list containing signature gene lists. At least two signature gene lists must be present.

expectedCountsMatrix

a normalized gene expression data matrix with rows corresponding to genes and columns to samples.

permRatio

a numeric between 0 and 1 representing the number of samples that are retained for the permutation step. The rounded value of the number of patients multiplied by this numeric should be inferior to the total number of patients so that the permutation step is not done on the entire cohort. Default: 0.75.

permNbr

a integer, 5 or higher, representing the number of permutation sampling done. In addition, the number of permutations must be equal or inferior to the total number of unique permutations with the dataset considering the parameters selected by the user. Default: 20.

upscaleNbr

a integer, 2 or higher, representing the number of values taken for the normal distribution for each sample to run the upscaling step. Default: 10.

Value

TRUE

Author

Astrid Deschênes

Examples

## Loading signatures
data("signatures")

## Load demo normalized expected counts for 30 patients
data("expNormalCountsDemo")

## Run subtyping on the 30 patients
runSubtyping(geneLists=signatures, expectedCountsMatrix=expNormalCountsDemo, 
    permRatio=0.75, permNbr=20, upscaleNbr=10)
#> $GSVA_RESULTS
#>                                            Patient_1  Patient_2  Patient_3
#> 2018_Tiriac_PDAC_PDO_classical_signature   0.1422663  0.4274225 -0.7532768
#> 2018_Tiriac_PDAC_PDO_basal-like_signature -0.2385417 -0.4403292  0.8006550
#>                                            Patient_4  Patient_5  Patient_6
#> 2018_Tiriac_PDAC_PDO_classical_signature   0.3479971 -0.6732937  0.5735090
#> 2018_Tiriac_PDAC_PDO_basal-like_signature -0.4500838  0.8175287 -0.6076389
#>                                            Patient_7  Patient_8  Patient_9
#> 2018_Tiriac_PDAC_PDO_classical_signature  -0.1247647 -0.1261704 -0.1287452
#> 2018_Tiriac_PDAC_PDO_basal-like_signature  0.1209839  0.2276770  0.1643817
#>                                           Patient_10 Patient_11 Patient_12
#> 2018_Tiriac_PDAC_PDO_classical_signature   0.4370278  0.5913501 -0.2002262
#> 2018_Tiriac_PDAC_PDO_basal-like_signature -0.4448441 -0.6485138  0.1993417
#>                                           Patient_13 Patient_14   Patient_15
#> 2018_Tiriac_PDAC_PDO_classical_signature   0.2448591  0.3249211  0.053730189
#> 2018_Tiriac_PDAC_PDO_basal-like_signature -0.2308458 -0.3733219 -0.005216285
#>                                           Patient_16 Patient_17 Patient_18
#> 2018_Tiriac_PDAC_PDO_classical_signature  -0.3930325 -0.7043936  0.7082490
#> 2018_Tiriac_PDAC_PDO_basal-like_signature  0.4381176  0.7356150 -0.8162717
#>                                            Patient_19 Patient_20 Patient_21
#> 2018_Tiriac_PDAC_PDO_classical_signature  -0.05232661 -0.4716993  0.4735043
#> 2018_Tiriac_PDAC_PDO_basal-like_signature  0.14282238  0.5886179 -0.6005646
#>                                           Patient_22 Patient_23 Patient_24
#> 2018_Tiriac_PDAC_PDO_classical_signature  -0.2313920  0.5079302  0.4728932
#> 2018_Tiriac_PDAC_PDO_basal-like_signature  0.2876812 -0.5792517 -0.4255245
#>                                           Patient_25 Patient_26 Patient_27
#> 2018_Tiriac_PDAC_PDO_classical_signature  -0.8846154 -0.5213023  0.7477403
#> 2018_Tiriac_PDAC_PDO_basal-like_signature  0.9387755  0.6288557 -0.7916667
#>                                           Patient_28 Patient_29 Patient_30
#> 2018_Tiriac_PDAC_PDO_classical_signature  -0.2808672  0.6611142  0.3823810
#> 2018_Tiriac_PDAC_PDO_basal-like_signature  0.3847667 -0.7322511 -0.4204301
#> attr(,"gsvaParam")
#> attr(,"gsvaParam")$originalClassWasSE
#> [1] FALSE
#> 
#> attr(,"gsvaParam")$geneSets
#> attr(,"gsvaParam")$geneSets$`2018_Tiriac_PDAC_PDO_classical_signature`
#>  [1] "MUC13"      "GPR35"      "USH1C"      "BAIAP2L2"   "GAL3ST1"   
#>  [6] "VIL1"       "DDC"        "LGALS4"     "TSPAN8"     "SLC5A1"    
#> [11] "SEMA4G"     "TNFRSF11A"  "HSD17B2"    "CKMT1A"     "DPCR1"     
#> [16] "GCNT3"      "SLC6A20"    "POF1B"      "NR1I2"      "MUC17"     
#> [21] "HHLA2"      "BTNL8"      "HEPH"       "PDZD3"      "MYO1A"     
#> [26] "CDH17"      "MYO7B"      "GPA33"      "CALML4"     "ANXA13"    
#> [31] "TM4SF20"    "SI"         "SLC44A4"    "FAM83E"     "SLC40A1"   
#> [36] "CTSE"       "BACE2"      "SULT1C2"    "AGR3"       "ATP7B"     
#> [41] "LYZ"        "TESC"       "TOX3"       "SH3BGRL2"   "GMDS"      
#> [46] "GATA6"      "GPX2"       "CA2"        "ST6GALNAC1" "FOXA3"     
#> [51] "AKR1C3"     "TMC5"       "CYP2S1"     "PAQR8"      "HKDC1"     
#> [56] "IYD"        "DOK4"       "CYP2C18"    "KCNE3"      "FZD5"      
#> [61] "RASEF"      "GKN1"      
#> 
#> attr(,"gsvaParam")$geneSets$`2018_Tiriac_PDAC_PDO_basal-like_signature`
#>  [1] "ANO1"     "ADORA2B"  "S100A2"   "CSF2"     "ANXA1"    "EREG"    
#>  [7] "RGS20"    "CELSR1"   "WNT7B"    "AMIGO2"   "PTHLH"    "C12orf54"
#> [13] "TGM2"     "REEP3"    "DCBLD2"   "NMB"      "IKZF2"    "GPR87"   
#> [19] "MUC16"    "IVL"      "SPRR1B"   "BCAR3"    "TTC9"     "PTGES"   
#> [25] "CST6"     "CATSPER1"
#> 
#> 
#> attr(,"gsvaParam")$assay
#> [1] NA
#> 
#> attr(,"gsvaParam")$annotation
#> geneIdType: Null
#> 
#> attr(,"gsvaParam")$minSize
#> [1] 1
#> 
#> attr(,"gsvaParam")$maxSize
#> [1] Inf
#> 
#> attr(,"gsvaParam")$nzcount
#> [1] 2580
#> 
#> attr(,"gsvaParam")$ondisk
#> [1] "auto"
#> 
#> attr(,"gsvaParam")$checkNA
#> [1] "auto"
#> 
#> attr(,"gsvaParam")$didCheckNA
#> [1] TRUE
#> 
#> attr(,"gsvaParam")$anyNA
#> [1] FALSE
#> 
#> attr(,"gsvaParam")$use
#> [1] "everything"
#> 
#> attr(,"gsvaParam")$kcdf
#> [1] "auto"
#> 
#> attr(,"gsvaParam")$kcdfNoneMinSampleSize
#> [1] 200
#> 
#> attr(,"gsvaParam")$tau
#> [1] 1
#> 
#> attr(,"gsvaParam")$maxDiff
#> [1] TRUE
#> 
#> attr(,"gsvaParam")$absRanking
#> [1] FALSE
#> 
#> attr(,"gsvaParam")$sparse
#> [1] FALSE
#> 
#> attr(,"gsvaParam")$filterRows
#> [1] TRUE
#> 
#> attr(,"assay")
#> [1] "es"
#> attr(,"geneSets")
#> attr(,"geneSets")$`2018_Tiriac_PDAC_PDO_classical_signature`
#>  [1] "MUC13"      "GPR35"      "USH1C"      "BAIAP2L2"   "GAL3ST1"   
#>  [6] "VIL1"       "DDC"        "LGALS4"     "TSPAN8"     "SLC5A1"    
#> [11] "SEMA4G"     "TNFRSF11A"  "HSD17B2"    "CKMT1A"     "GCNT3"     
#> [16] "SLC6A20"    "POF1B"      "NR1I2"      "MUC17"      "HHLA2"     
#> [21] "BTNL8"      "HEPH"       "MYO1A"      "CDH17"      "MYO7B"     
#> [26] "GPA33"      "CALML4"     "ANXA13"     "TM4SF20"    "SI"        
#> [31] "SLC44A4"    "FAM83E"     "SLC40A1"    "CTSE"       "BACE2"     
#> [36] "SULT1C2"    "AGR3"       "ATP7B"      "LYZ"        "TESC"      
#> [41] "TOX3"       "SH3BGRL2"   "GMDS"       "GATA6"      "GPX2"      
#> [46] "CA2"        "ST6GALNAC1" "FOXA3"      "AKR1C3"     "TMC5"      
#> [51] "CYP2S1"     "PAQR8"      "HKDC1"      "IYD"        "DOK4"      
#> [56] "CYP2C18"    "KCNE3"      "FZD5"       "RASEF"      "GKN1"      
#> 
#> attr(,"geneSets")$`2018_Tiriac_PDAC_PDO_basal-like_signature`
#>  [1] "ANO1"     "ADORA2B"  "S100A2"   "CSF2"     "ANXA1"    "EREG"    
#>  [7] "RGS20"    "CELSR1"   "WNT7B"    "AMIGO2"   "PTHLH"    "C12orf54"
#> [13] "TGM2"     "REEP3"    "DCBLD2"   "NMB"      "IKZF2"    "GPR87"   
#> [19] "MUC16"    "IVL"      "SPRR1B"   "BCAR3"    "TTC9"     "PTGES"   
#> [25] "CST6"     "CATSPER1"
#> 
#> 
#> $PERMUTATIONS
#> $PERMUTATIONS$`2018_Tiriac_PDAC_PDO_classical_signature`
#>                    [,1]         [,2]        [,3]        [,4]        [,5]
#> Patient_1   0.131724485           NA  0.22069141  0.23843619          NA
#> Patient_2            NA  0.372704531  0.47622785          NA          NA
#> Patient_3            NA -0.759967781 -0.74805532          NA -0.74658703
#> Patient_4            NA           NA          NA  0.32853324  0.40674722
#> Patient_5            NA -0.676976403 -0.62890808 -0.65397686          NA
#> Patient_6            NA  0.479231378  0.56648183  0.56112853  0.55661538
#> Patient_7  -0.158513035 -0.217343024          NA -0.15259484 -0.13456653
#> Patient_8  -0.179083771 -0.317989580 -0.09858131 -0.22437907 -0.22048508
#> Patient_9  -0.175573158 -0.112124185          NA -0.08070433 -0.09621108
#> Patient_10  0.416999345           NA  0.38000000  0.45746815  0.45191576
#> Patient_11  0.575321821  0.551513811  0.59241239          NA  0.57052338
#> Patient_12 -0.239263070 -0.302986573          NA -0.12650639 -0.22445913
#> Patient_13  0.166397369  0.189543637          NA  0.20313667  0.16400092
#> Patient_14  0.295686275  0.319930070  0.30012019  0.32858873          NA
#> Patient_15  0.007030604  0.005532596  0.09741671  0.08347386  0.03962302
#> Patient_16 -0.412009478 -0.385162231          NA          NA -0.33762631
#> Patient_17           NA           NA -0.68498507 -0.67856125          NA
#> Patient_18           NA  0.650274951  0.71031097          NA          NA
#> Patient_19 -0.144087256           NA -0.08648633 -0.04623155          NA
#> Patient_20 -0.487937350           NA          NA -0.46436789 -0.48678660
#> Patient_21  0.443776936  0.448729010  0.48277648  0.51537153  0.45764909
#> Patient_22 -0.219089193 -0.261019879 -0.21021322 -0.17448698 -0.22935564
#> Patient_23  0.515190181  0.405610311  0.53760206  0.54453756  0.51889972
#> Patient_24           NA  0.294149512  0.39431108  0.46273178  0.44574067
#> Patient_25 -0.884615385           NA -0.88461538          NA -0.92307692
#> Patient_26 -0.537591076 -0.563811189 -0.57682795 -0.47578099 -0.47501010
#> Patient_27  0.766255144  0.738524590          NA          NA  0.74651354
#> Patient_28 -0.401804078 -0.296191554 -0.29681398 -0.30531001 -0.34061587
#> Patient_29  0.670105315           NA  0.65536687  0.68885765  0.69218362
#> Patient_30  0.341493629  0.321080818  0.39190092          NA          NA
#>                  [,6]        [,7]        [,8]         [,9]       [,10]
#> Patient_1          NA  0.33097411  0.07356299           NA  0.24683425
#> Patient_2   0.3832970  0.51550960  0.42235123           NA  0.48191882
#> Patient_3  -0.7708530 -0.70967742          NA           NA -0.74117647
#> Patient_4          NA  0.48320351          NA  0.432560590  0.40191225
#> Patient_5  -0.6888112 -0.60090119 -0.69217309 -0.665816327 -0.69047619
#> Patient_6          NA  0.63293384  0.49850070           NA  0.59574724
#> Patient_7  -0.2171150  0.06623058 -0.21436939 -0.017208167 -0.06139307
#> Patient_8  -0.2071721          NA          NA -0.001010581 -0.01494793
#> Patient_9  -0.1033581          NA -0.18563492           NA  0.05635948
#> Patient_10  0.3486008  0.50573523          NA  0.473563287  0.46384615
#> Patient_11  0.5128414  0.65991903  0.57567331  0.619736927  0.59577429
#> Patient_12 -0.2964429 -0.07991038 -0.25722104 -0.196315349 -0.05980335
#> Patient_13  0.2114044  0.41996377          NA           NA          NA
#> Patient_14  0.2868590  0.36252679  0.29677380  0.359810260          NA
#> Patient_15  0.1162474  0.23622683 -0.03702494  0.117586118          NA
#> Patient_16 -0.3951705 -0.24632736 -0.45036462 -0.310487905          NA
#> Patient_17         NA -0.72099564          NA -0.705716302 -0.72465195
#> Patient_18  0.6683054          NA  0.69984855  0.722408027          NA
#> Patient_19         NA  0.15378652 -0.15977477  0.075310559          NA
#> Patient_20 -0.5110703          NA -0.51451020 -0.453785345 -0.44798184
#> Patient_21  0.4417395  0.60057194  0.39455488  0.502231557  0.52514916
#> Patient_22 -0.2585279 -0.10926311 -0.25934579 -0.170771610 -0.28561873
#> Patient_23  0.4324708          NA  0.48677885  0.537077804          NA
#> Patient_24         NA  0.54314766  0.35255413           NA  0.45630477
#> Patient_25         NA -0.84491784 -0.88461538 -0.884615385 -0.90404743
#> Patient_26 -0.5644054 -0.39971958 -0.56198025 -0.503415257 -0.43503833
#> Patient_27  0.7361111          NA  0.74271845  0.748960931  0.76342282
#> Patient_28         NA -0.16794037          NA -0.175518176          NA
#> Patient_29  0.6346935          NA  0.61225439           NA  0.70886356
#> Patient_30  0.3164152          NA          NA  0.414642115  0.45102163
#>                  [,11]       [,12]       [,13]      [,14]       [,15]
#> Patient_1   0.18852823  0.07692308  0.20456467  0.1879987  0.10982470
#> Patient_2           NA  0.33717949  0.45701357  0.4998587          NA
#> Patient_3  -0.73941227 -0.79264957 -0.74343501 -0.7232406          NA
#> Patient_4   0.38720830          NA          NA  0.4098356          NA
#> Patient_5  -0.67001212 -0.69849691 -0.65658616 -0.6704996 -0.70120898
#> Patient_6   0.57272343  0.44682378  0.53938388  0.5949519  0.52628863
#> Patient_7           NA          NA          NA -0.1696434 -0.10617735
#> Patient_8   0.00499304 -0.31788244          NA -0.0703808          NA
#> Patient_9  -0.04283283 -0.15381655 -0.07349581         NA          NA
#> Patient_10          NA  0.32668988  0.42692919         NA  0.46046010
#> Patient_11  0.60422039  0.51022552          NA  0.6032088  0.60744010
#> Patient_12          NA -0.31266223 -0.15035532 -0.1801154 -0.23447781
#> Patient_13  0.33604162  0.20981561  0.26905216  0.2813007  0.20931611
#> Patient_14          NA  0.27778482  0.36613070  0.2862090  0.28227692
#> Patient_15          NA          NA  0.12777053  0.1077561  0.06278327
#> Patient_16          NA -0.40016103 -0.35709135         NA          NA
#> Patient_17 -0.68973890          NA -0.64314257 -0.6813337 -0.68025100
#> Patient_18  0.71031097  0.61532046  0.68949816  0.7228927  0.69288175
#> Patient_19 -0.10097273 -0.28960170 -0.05864009 -0.1008983 -0.17141574
#> Patient_20 -0.46369081 -0.50918876 -0.47836538         NA -0.51314898
#> Patient_21  0.50111673          NA          NA         NA          NA
#> Patient_22 -0.25718726          NA -0.16437183         NA -0.26120556
#> Patient_23  0.50937041  0.44287772  0.47891417  0.5074191  0.50588163
#> Patient_24          NA  0.26832862  0.44727478         NA  0.40157668
#> Patient_25 -0.88461538          NA          NA -0.9042090 -0.90429758
#> Patient_26 -0.50563660 -0.60215924 -0.51647425 -0.5334448 -0.52523937
#> Patient_27  0.75331126  0.72441663          NA  0.7638315  0.76910435
#> Patient_28 -0.29009754          NA          NA -0.2673299          NA
#> Patient_29  0.68118166  0.58620039  0.67181642         NA  0.62337421
#> Patient_30  0.39793827  0.24020896  0.38784119  0.4120185  0.35403876
#>                 [,16]       [,17]       [,18]       [,19]       [,20]
#> Patient_1          NA  0.17220280          NA          NA          NA
#> Patient_2          NA  0.43016713  0.33681603  0.40387569  0.43016791
#> Patient_3          NA -0.75609509          NA -0.74910323 -0.77031039
#> Patient_4   0.2882639  0.39823718  0.26138148  0.35973036          NA
#> Patient_5  -0.6981943          NA -0.72217429 -0.67112887 -0.68021978
#> Patient_6   0.5399782  0.56553943          NA  0.49186212          NA
#> Patient_7  -0.2330163          NA -0.24811066 -0.23080230          NA
#> Patient_8  -0.1831502 -0.18902825 -0.30976801 -0.18729097 -0.12211241
#> Patient_9  -0.1235201 -0.16371681 -0.14047196 -0.05079272 -0.10347099
#> Patient_10  0.3873626  0.43639397          NA  0.46653816  0.43883297
#> Patient_11  0.5363195          NA  0.54665416          NA          NA
#> Patient_12         NA -0.27686345 -0.24016136          NA -0.24530446
#> Patient_13  0.2124278  0.23848492  0.07148702  0.17769693  0.28785680
#> Patient_14         NA  0.34183232  0.30830524  0.32403539  0.34770785
#> Patient_15         NA  0.06802024 -0.01076658  0.02777778  0.07139644
#> Patient_16 -0.4456823 -0.41443190          NA -0.32206682 -0.37398824
#> Patient_17 -0.6743307          NA -0.70768520 -0.66755975 -0.69623675
#> Patient_18  0.6362939          NA  0.66202946          NA  0.67839196
#> Patient_19 -0.0762651 -0.02599551 -0.27096433          NA -0.05289207
#> Patient_20 -0.5221029          NA          NA          NA          NA
#> Patient_21  0.4136578  0.49323753  0.42186448  0.49492080  0.45656927
#> Patient_22 -0.3360528 -0.26025937 -0.28993131 -0.22452547          NA
#> Patient_23  0.4356468  0.51600269  0.43197926  0.51153846  0.53258764
#> Patient_24  0.4054666          NA  0.25007637          NA  0.44171415
#> Patient_25 -0.8846154 -0.88461538          NA          NA -0.88645720
#> Patient_26         NA -0.52785597 -0.60412758 -0.54081350          NA
#> Patient_27  0.7446809  0.75695581  0.74529075  0.76119403  0.75362319
#> Patient_28         NA -0.32100006 -0.34715285 -0.30744215 -0.28366149
#> Patient_29  0.6416987          NA  0.64090266  0.67154472  0.72204214
#> Patient_30  0.3247732  0.36925477          NA  0.33437615  0.39780395
#> 
#> $PERMUTATIONS$`2018_Tiriac_PDAC_PDO_basal-like_signature`
#>                   [,1]        [,2]        [,3]       [,4]         [,5]
#> Patient_1  -0.16800000          NA -0.29660446 -0.2573491           NA
#> Patient_2           NA -0.36034946 -0.47421384         NA           NA
#> Patient_3           NA  0.82785680  0.79161850         NA  0.798941408
#> Patient_4           NA          NA          NA -0.4410234 -0.430284192
#> Patient_5           NA  0.84700240  0.76761905  0.8211310           NA
#> Patient_6           NA -0.52275313 -0.58042813 -0.5983421 -0.611324430
#> Patient_7   0.15198777  0.24504249          NA  0.1574156  0.123552894
#> Patient_8   0.25494505  0.35757576  0.19385298  0.2868237  0.249549550
#> Patient_9   0.19007092  0.09069767          NA  0.1020576  0.112376238
#> Patient_10 -0.47706093          NA -0.42604736 -0.5103976 -0.442509363
#> Patient_11 -0.68397436 -0.61972574 -0.66166667         NA -0.680183727
#> Patient_12  0.25757845  0.21535581          NA  0.2127208  0.186584651
#> Patient_13 -0.19496124 -0.20135551          NA -0.2588235 -0.287630208
#> Patient_14 -0.35952381 -0.34292274 -0.33971880 -0.4267703           NA
#> Patient_15  0.05208333 -0.01342926 -0.02260062 -0.1324284  0.001449275
#> Patient_16  0.49375804  0.46675243          NA         NA  0.474812030
#> Patient_17          NA          NA  0.74037375  0.7529259           NA
#> Patient_18          NA -0.79567511 -0.81921850         NA           NA
#> Patient_19  0.21683039          NA  0.15609319  0.1913998           NA
#> Patient_20  0.59234020          NA          NA  0.6183007  0.610426929
#> Patient_21 -0.58758073 -0.63333333 -0.60504286 -0.6663943 -0.585741811
#> Patient_22  0.31100917  0.25887231  0.27863720  0.1578483  0.270626151
#> Patient_23 -0.57384220 -0.40566038 -0.60984340 -0.5709220 -0.602218330
#> Patient_24          NA -0.27466433 -0.40551378 -0.3832207 -0.410663236
#> Patient_25  0.95213454          NA  0.94723295         NA  0.963057325
#> Patient_26  0.62642276  0.69397163  0.64119497  0.5925189  0.598231966
#> Patient_27 -0.79411146 -0.76261261          NA         NA -0.792857143
#> Patient_28  0.43888889  0.44400000  0.36217105  0.3966238  0.361965812
#> Patient_29 -0.72500000          NA -0.74921260 -0.7701550 -0.724958949
#> Patient_30 -0.40443425 -0.29743590 -0.41271478         NA           NA
#>                   [,6]        [,7]       [,8]         [,9]       [,10]
#> Patient_1           NA -0.36311475 -0.1900498           NA -0.22229299
#> Patient_2  -0.41301775 -0.47676768 -0.3562633           NA -0.47226721
#> Patient_3   0.82728551  0.79479167         NA           NA  0.79529590
#> Patient_4           NA -0.55470085         NA -0.465021930 -0.50000000
#> Patient_5   0.86063830  0.77339450  0.8252597  0.842372462  0.81562655
#> Patient_6           NA -0.69365482 -0.5894144           NA -0.64172494
#> Patient_7   0.24008439 -0.02087327  0.1726887  0.027726575  0.05340633
#> Patient_8   0.31515651          NA         NA  0.100450450  0.13684211
#> Patient_9   0.19476190          NA  0.1859025           NA  0.03420398
#> Patient_10 -0.49007491 -0.58625304         NA -0.528571429 -0.48324226
#> Patient_11 -0.63743961 -0.74089980 -0.6758900 -0.705284553 -0.65932070
#> Patient_12  0.28013333 -0.01960784  0.2498858  0.097720545  0.20895954
#> Patient_13 -0.15232975 -0.41530612         NA           NA          NA
#> Patient_14 -0.35618414 -0.45685218 -0.3188815 -0.420138889          NA
#> Patient_15 -0.08913043 -0.18120567  0.0212618 -0.102773575          NA
#> Patient_16  0.45789474  0.35125241  0.5246149  0.436654367          NA
#> Patient_17          NA  0.75496689         NA  0.762845850  0.74757174
#> Patient_18 -0.79049080          NA -0.7785714 -0.829555896          NA
#> Patient_19          NA -0.08464520  0.2650538  0.007040998          NA
#> Patient_20  0.65939968          NA  0.6123360  0.560844749  0.60658263
#> Patient_21 -0.56256831 -0.71676892 -0.5749518 -0.641703594 -0.66769231
#> Patient_22  0.31344086  0.15157963  0.3690265  0.168585132  0.33047945
#> Patient_23 -0.45533211          NA -0.4765671 -0.600106496          NA
#> Patient_24          NA -0.57475490 -0.2900829           NA -0.47649573
#> Patient_25          NA  0.89334176  0.9399745  0.950904393  0.92944862
#> Patient_26  0.72686497  0.49813084  0.6926065  0.566666667  0.57116788
#> Patient_27 -0.77186667          NA -0.7965198 -0.796542828 -0.82252664
#> Patient_28          NA  0.18639291         NA  0.333014862          NA
#> Patient_29 -0.71132686          NA -0.6649123           NA -0.74500526
#> Patient_30 -0.32513661          NA         NA -0.496089049 -0.34100762
#>                   [,11]       [,12]       [,13]       [,14]        [,15]
#> Patient_1  -0.237669801 -0.12065112 -0.26876877 -0.25662526 -0.168746904
#> Patient_2            NA -0.37826087 -0.42664117 -0.45628415           NA
#> Patient_3   0.798651252  0.85156538  0.81081277  0.79465875           NA
#> Patient_4  -0.403731343          NA          NA -0.53000000           NA
#> Patient_5   0.815134371  0.82887651  0.77921348  0.80948905  0.835576459
#> Patient_6  -0.642708333 -0.51716992 -0.58773690 -0.61818981 -0.602405498
#> Patient_7            NA          NA          NA  0.07205285  0.146448087
#> Patient_8   0.100922722  0.33030551          NA  0.07249071           NA
#> Patient_9   0.004089219  0.21030303  0.06442786          NA           NA
#> Patient_10           NA -0.44084967 -0.51954315          NA -0.475516224
#> Patient_11 -0.651916933 -0.59337748          NA -0.66320755 -0.691070482
#> Patient_12           NA  0.33133544  0.25609756  0.06289134  0.287625814
#> Patient_13 -0.284271420 -0.09234973 -0.28243567 -0.36384977 -0.158045977
#> Patient_14           NA -0.32979003 -0.43321957 -0.38670034 -0.303978854
#> Patient_15           NA          NA -0.22892377 -0.01200828 -0.001863354
#> Patient_16           NA  0.49173531  0.41693029          NA           NA
#> Patient_17  0.746560847          NA  0.74228275  0.75062057  0.763852243
#> Patient_18 -0.791706539 -0.75205128 -0.82705167 -0.84341693 -0.785355106
#> Patient_19  0.083928571  0.34360731  0.13747885  0.15107914  0.252836879
#> Patient_20  0.561471861  0.62946824  0.62409163          NA  0.591666667
#> Patient_21 -0.599262537          NA          NA          NA           NA
#> Patient_22  0.260379597          NA  0.20405405          NA  0.305855856
#> Patient_23 -0.578302417 -0.29929453 -0.55564738 -0.55273224 -0.515000000
#> Patient_24           NA -0.25350877 -0.42536232          NA -0.382620321
#> Patient_25  0.934010152          NA          NA  0.94024081  0.946343537
#> Patient_26  0.589356555  0.73843441  0.62956259  0.59596733  0.607884232
#> Patient_27 -0.796463859 -0.79387646          NA -0.79811912 -0.816772152
#> Patient_28  0.335903814          NA          NA  0.37440945           NA
#> Patient_29 -0.742724458 -0.69962726 -0.74835610          NA -0.699697968
#> Patient_30 -0.345287958 -0.25960452 -0.31379310 -0.43027444 -0.288674839
#>                 [,16]       [,17]       [,18]       [,19]        [,20]
#> Patient_1          NA -0.20213398          NA          NA           NA
#> Patient_2          NA -0.39951591 -0.30060241 -0.39865935 -0.433739013
#> Patient_3          NA  0.79061033          NA  0.82624216  0.813328632
#> Patient_4  -0.3756849 -0.45929007 -0.29028986 -0.44893843           NA
#> Patient_5   0.8460641          NA  0.85942029  0.84543708  0.814403292
#> Patient_6  -0.5780551 -0.58487973          NA -0.55264026           NA
#> Patient_7   0.2104790          NA  0.26824165  0.18736638           NA
#> Patient_8   0.3180751  0.25989583  0.31627051  0.30469522  0.208258528
#> Patient_9   0.2044025  0.22437458  0.12944664  0.10392703  0.141567852
#> Patient_10 -0.4372420 -0.48100358          NA -0.52167832 -0.477138643
#> Patient_11 -0.5748194          NA -0.58896069          NA           NA
#> Patient_12         NA  0.26219439  0.28000000          NA  0.158111308
#> Patient_13 -0.1823899 -0.28461538 -0.08489681 -0.27540297 -0.259994223
#> Patient_14         NA -0.33051410 -0.30993266 -0.41926298 -0.383274021
#> Patient_15         NA  0.01018519  0.03081395  0.00131516  0.006134969
#> Patient_16  0.5095588  0.46155268          NA  0.45172005  0.444318182
#> Patient_17  0.7644737          NA  0.80648390  0.77620932  0.750596760
#> Patient_18 -0.7521156          NA -0.79776715          NA -0.803944020
#> Patient_19  0.2319124  0.17979890  0.43333333          NA  0.152534287
#> Patient_20  0.6333063          NA          NA          NA           NA
#> Patient_21 -0.5759003 -0.59219219 -0.55617633 -0.64078319 -0.612306202
#> Patient_22  0.3553381  0.31636364  0.27970297  0.25958941           NA
#> Patient_23 -0.4771930 -0.54132883 -0.45285344 -0.44453552 -0.561488673
#> Patient_24 -0.3495770          NA -0.29689266          NA -0.365220012
#> Patient_25  0.9435897  0.94723295          NA          NA  0.938775510
#> Patient_26         NA  0.64287834  0.72220637  0.68308458           NA
#> Patient_27 -0.8074960 -0.79545455 -0.77526539 -0.82869362 -0.783717738
#> Patient_28         NA  0.37770193  0.42814695  0.40786214  0.380118443
#> Patient_29 -0.7174603          NA -0.69944259 -0.75417340 -0.802032520
#> Patient_30 -0.3114180 -0.39691358          NA -0.35770308 -0.472408964
#> 
#> 
#> $SD
#> $SD$`2018_Tiriac_PDAC_PDO_classical_signature`
#>  Patient_1  Patient_2  Patient_3  Patient_4  Patient_5  Patient_6  Patient_7 
#> 0.07525652 0.05853536 0.02120160 0.06433028 0.02882614 0.04942779 0.09244187 
#>  Patient_8  Patient_9 Patient_10 Patient_11 Patient_12 Patient_13 Patient_14 
#> 0.10684281 0.06177529 0.04941508 0.04096201 0.07660467 0.07956310 0.03042514 
#> Patient_15 Patient_16 Patient_17 Patient_18 Patient_19 Patient_20 Patient_21 
#> 0.06662980 0.05768925 0.02238712 0.03358856 0.11419969 0.02610208 0.05072414 
#> Patient_22 Patient_23 Patient_24 Patient_25 Patient_26 Patient_27 Patient_28 
#> 0.05533580 0.04329473 0.08509572 0.01811274 0.05392747 0.01232988 0.06350747 
#> Patient_29 Patient_30 
#> 0.03718490 0.05248334 
#> 
#> $SD$`2018_Tiriac_PDAC_PDO_basal-like_signature`
#>  Patient_1  Patient_2  Patient_3  Patient_4  Patient_5  Patient_6  Patient_7 
#> 0.06547320 0.05322573 0.01884877 0.07319749 0.02822649 0.04537055 0.08625883 
#>  Patient_8  Patient_9 Patient_10 Patient_11 Patient_12 Patient_13 Patient_14 
#> 0.09209113 0.06786416 0.04231860 0.04582427 0.09289786 0.09054065 0.04889792 
#> Patient_15 Patient_16 Patient_17 Patient_18 Patient_19 Patient_20 Patient_21 
#> 0.08126758 0.04433780 0.01760890 0.02774835 0.12558308 0.02865780 0.04418419 
#> Patient_22 Patient_23 Patient_24 Patient_25 Patient_26 Patient_27 Patient_28 
#> 0.06589751 0.08194343 0.08847561 0.01652843 0.06463532 0.01758221 0.06599959 
#> Patient_29 Patient_30 
#> 0.03385338 0.06938392 
#> 
#>