A list of published gene signatures, mostly related to PDAC classification.

A list of published gene signatures, mostly related to PDAC classification.

data(signatures)

Format

a list containing one entry per signature:

"2018_Tiriac_PDAC_PDO_classical_signature": a vector of gene names associated to the PDAC Patient-derived Organoid (PDO) Classical signature as published in Tiriac et al 2018

"2018_Tiriac_PDAC_PDO_basal-like_signature": a vector of gene names associated to the PDAC Patient-derived Organoid (PDO) Basal-like signature as published in Tiriac et al 2018

Value

a list containing one entry per signature:

"2018_Tiriac_PDAC_PDO_classical_signature": a vector of gene names associated to the PDAC Patient-derived Organoid (PDO) Classical signature as published in Tiriac et al 2018

"2018_Tiriac_PDAC_PDO_basal-like_signature": a vector of gene names associated to the PDAC Patient-derived Organoid (PDO) Basal-like signature as published in Tiriac et al 2018

Details

The PDAC PDO classical and basal-like signatures are associated to this publication:

Tiriac et al. Organoid Profiling Identifies Common Responders to Chemotherapy in Pancreatic Cancer. Cancer Discov. 2018 Sep;8(9):1112-1129. doi: 10.1158/2159-8290.CD-18-0349. Epub 2018 May 31. PMID: 29853643; PMCID: PMC6125219.

See also

Examples

## Load the published signature gene list
data(signatures)