TODO

TODO.

runSubtyping(
  geneLists,
  expectedCountsMatrix,
  permRatio = 0.75,
  permNbr = 20,
  upscaleNbr = 10
)

Arguments

geneLists

a list containing signature gene lists. At least two signature gene lists must be present.

expectedCountsMatrix

a normalized gene expression data matrix with rows corresponding to genes and columns to samples.

permRatio

a numeric between 0 and 1 representing the number of samples that are retained for the permutation step. The rounded value of the number of patients multiplied by this numeric should be inferior to the total number of patients so that the permutation step is not done on the entire cohort. Default: 0.75.

permNbr

a integer, 5 or higher, representing the number of permutation sampling done. In addition, the number of permutations must be equal or inferior to the total number of unique permutations with the dataset considering the parameters selected by the user. Default: 20.

upscaleNbr

a integer, 2 or higher, representing the number of values taken for the normal distribution for each sample to run the upscaling step. Default: 10.

Value

TRUE

Author

Astrid Deschênes

Examples

## Loading signatures
data("signaturesDemo")

## Load demo normalized expected counts for 30 patients
data("expNormalCountsDemo")

## Run subtyping on the 30 patients
runSubtyping(geneLists=signaturesDemo, 
    expectedCountsMatrix=expNormalCountsDemo, 
    permRatio=0.75, permNbr=20, upscaleNbr=10)
#> number of iterations= 37 
#> number of iterations= 24 
#> $GSVA_RESULTS
#>                                            Patient_1  Patient_2  Patient_3
#> 2018_Tiriac_PDAC_PDO_classical_signature   0.1422663  0.4274225 -0.7532768
#> 2018_Tiriac_PDAC_PDO_basal-like_signature -0.2385417 -0.4403292  0.8006550
#>                                            Patient_4  Patient_5  Patient_6
#> 2018_Tiriac_PDAC_PDO_classical_signature   0.3479971 -0.6732937  0.5735090
#> 2018_Tiriac_PDAC_PDO_basal-like_signature -0.4500838  0.8175287 -0.6076389
#>                                            Patient_7  Patient_8  Patient_9
#> 2018_Tiriac_PDAC_PDO_classical_signature  -0.1247647 -0.1261704 -0.1287452
#> 2018_Tiriac_PDAC_PDO_basal-like_signature  0.1209839  0.2276770  0.1643817
#>                                           Patient_10 Patient_11 Patient_12
#> 2018_Tiriac_PDAC_PDO_classical_signature   0.4370278  0.5913501 -0.2002262
#> 2018_Tiriac_PDAC_PDO_basal-like_signature -0.4448441 -0.6485138  0.1993417
#>                                           Patient_13 Patient_14   Patient_15
#> 2018_Tiriac_PDAC_PDO_classical_signature   0.2448591  0.3249211  0.053730189
#> 2018_Tiriac_PDAC_PDO_basal-like_signature -0.2308458 -0.3733219 -0.005216285
#>                                           Patient_16 Patient_17 Patient_18
#> 2018_Tiriac_PDAC_PDO_classical_signature  -0.3930325 -0.7043936  0.7082490
#> 2018_Tiriac_PDAC_PDO_basal-like_signature  0.4381176  0.7356150 -0.8162717
#>                                            Patient_19 Patient_20 Patient_21
#> 2018_Tiriac_PDAC_PDO_classical_signature  -0.05232661 -0.4716993  0.4735043
#> 2018_Tiriac_PDAC_PDO_basal-like_signature  0.14282238  0.5886179 -0.6005646
#>                                           Patient_22 Patient_23 Patient_24
#> 2018_Tiriac_PDAC_PDO_classical_signature  -0.2313920  0.5079302  0.4728932
#> 2018_Tiriac_PDAC_PDO_basal-like_signature  0.2876812 -0.5792517 -0.4255245
#>                                           Patient_25 Patient_26 Patient_27
#> 2018_Tiriac_PDAC_PDO_classical_signature  -0.8846154 -0.5213023  0.7477403
#> 2018_Tiriac_PDAC_PDO_basal-like_signature  0.9387755  0.6288557 -0.7916667
#>                                           Patient_28 Patient_29 Patient_30
#> 2018_Tiriac_PDAC_PDO_classical_signature  -0.2808672  0.6611142  0.3823810
#> 2018_Tiriac_PDAC_PDO_basal-like_signature  0.3847667 -0.7322511 -0.4204301
#> attr(,"gsvaParam")
#> attr(,"gsvaParam")$originalClassWasSE
#> [1] FALSE
#> 
#> attr(,"gsvaParam")$geneSets
#> attr(,"gsvaParam")$geneSets$`2018_Tiriac_PDAC_PDO_classical_signature`
#>  [1] "MUC13"      "GPR35"      "USH1C"      "BAIAP2L2"   "GAL3ST1"   
#>  [6] "VIL1"       "DDC"        "LGALS4"     "TSPAN8"     "SLC5A1"    
#> [11] "SEMA4G"     "TNFRSF11A"  "HSD17B2"    "CKMT1A"     "DPCR1"     
#> [16] "GCNT3"      "SLC6A20"    "POF1B"      "NR1I2"      "MUC17"     
#> [21] "HHLA2"      "BTNL8"      "HEPH"       "PDZD3"      "MYO1A"     
#> [26] "CDH17"      "MYO7B"      "GPA33"      "CALML4"     "ANXA13"    
#> [31] "TM4SF20"    "SI"         "SLC44A4"    "FAM83E"     "SLC40A1"   
#> [36] "CTSE"       "BACE2"      "SULT1C2"    "AGR3"       "ATP7B"     
#> [41] "LYZ"        "TESC"       "TOX3"       "SH3BGRL2"   "GMDS"      
#> [46] "GATA6"      "GPX2"       "CA2"        "ST6GALNAC1" "FOXA3"     
#> [51] "AKR1C3"     "TMC5"       "CYP2S1"     "PAQR8"      "HKDC1"     
#> [56] "IYD"        "DOK4"       "CYP2C18"    "KCNE3"      "FZD5"      
#> [61] "RASEF"      "GKN1"      
#> 
#> attr(,"gsvaParam")$geneSets$`2018_Tiriac_PDAC_PDO_basal-like_signature`
#>  [1] "ANO1"     "ADORA2B"  "S100A2"   "CSF2"     "ANXA1"    "EREG"    
#>  [7] "RGS20"    "CELSR1"   "WNT7B"    "AMIGO2"   "PTHLH"    "C12orf54"
#> [13] "TGM2"     "REEP3"    "DCBLD2"   "NMB"      "IKZF2"    "GPR87"   
#> [19] "MUC16"    "IVL"      "SPRR1B"   "BCAR3"    "TTC9"     "PTGES"   
#> [25] "CST6"     "CATSPER1"
#> 
#> 
#> attr(,"gsvaParam")$assay
#> [1] NA
#> 
#> attr(,"gsvaParam")$annotation
#> geneIdType: Null
#> 
#> attr(,"gsvaParam")$minSize
#> [1] 1
#> 
#> attr(,"gsvaParam")$maxSize
#> [1] Inf
#> 
#> attr(,"gsvaParam")$nzcount
#> [1] 2580
#> 
#> attr(,"gsvaParam")$ondisk
#> [1] "auto"
#> 
#> attr(,"gsvaParam")$checkNA
#> [1] "auto"
#> 
#> attr(,"gsvaParam")$didCheckNA
#> [1] TRUE
#> 
#> attr(,"gsvaParam")$anyNA
#> [1] FALSE
#> 
#> attr(,"gsvaParam")$use
#> [1] "everything"
#> 
#> attr(,"gsvaParam")$kcdf
#> [1] "auto"
#> 
#> attr(,"gsvaParam")$kcdfNoneMinSampleSize
#> [1] 200
#> 
#> attr(,"gsvaParam")$tau
#> [1] 1
#> 
#> attr(,"gsvaParam")$maxDiff
#> [1] TRUE
#> 
#> attr(,"gsvaParam")$absRanking
#> [1] FALSE
#> 
#> attr(,"gsvaParam")$sparse
#> [1] FALSE
#> 
#> attr(,"gsvaParam")$filterRows
#> [1] TRUE
#> 
#> attr(,"assay")
#> [1] "es"
#> attr(,"geneSets")
#> attr(,"geneSets")$`2018_Tiriac_PDAC_PDO_classical_signature`
#>  [1] "MUC13"      "GPR35"      "USH1C"      "BAIAP2L2"   "GAL3ST1"   
#>  [6] "VIL1"       "DDC"        "LGALS4"     "TSPAN8"     "SLC5A1"    
#> [11] "SEMA4G"     "TNFRSF11A"  "HSD17B2"    "CKMT1A"     "GCNT3"     
#> [16] "SLC6A20"    "POF1B"      "NR1I2"      "MUC17"      "HHLA2"     
#> [21] "BTNL8"      "HEPH"       "MYO1A"      "CDH17"      "MYO7B"     
#> [26] "GPA33"      "CALML4"     "ANXA13"     "TM4SF20"    "SI"        
#> [31] "SLC44A4"    "FAM83E"     "SLC40A1"    "CTSE"       "BACE2"     
#> [36] "SULT1C2"    "AGR3"       "ATP7B"      "LYZ"        "TESC"      
#> [41] "TOX3"       "SH3BGRL2"   "GMDS"       "GATA6"      "GPX2"      
#> [46] "CA2"        "ST6GALNAC1" "FOXA3"      "AKR1C3"     "TMC5"      
#> [51] "CYP2S1"     "PAQR8"      "HKDC1"      "IYD"        "DOK4"      
#> [56] "CYP2C18"    "KCNE3"      "FZD5"       "RASEF"      "GKN1"      
#> 
#> attr(,"geneSets")$`2018_Tiriac_PDAC_PDO_basal-like_signature`
#>  [1] "ANO1"     "ADORA2B"  "S100A2"   "CSF2"     "ANXA1"    "EREG"    
#>  [7] "RGS20"    "CELSR1"   "WNT7B"    "AMIGO2"   "PTHLH"    "C12orf54"
#> [13] "TGM2"     "REEP3"    "DCBLD2"   "NMB"      "IKZF2"    "GPR87"   
#> [19] "MUC16"    "IVL"      "SPRR1B"   "BCAR3"    "TTC9"     "PTGES"   
#> [25] "CST6"     "CATSPER1"
#> 
#> 
#> $PERMUTATIONS
#> $PERMUTATIONS$`2018_Tiriac_PDAC_PDO_classical_signature`
#>                   [,1]        [,2]        [,3]        [,4]        [,5]
#> Patient_1   0.06542660  0.21742286  0.12155421          NA  0.19783791
#> Patient_2   0.38663071  0.48612213  0.31121463  0.44340901          NA
#> Patient_3           NA -0.72413453 -0.75962361          NA          NA
#> Patient_4   0.35628338          NA  0.32269262  0.42535021  0.35084578
#> Patient_5  -0.68830476 -0.62921727          NA -0.69665494 -0.67178159
#> Patient_6   0.54373938  0.57647271  0.51549509  0.55037494          NA
#> Patient_7  -0.22783518          NA -0.21880342 -0.14974619          NA
#> Patient_8           NA          NA -0.26313785          NA -0.16774898
#> Patient_9           NA -0.04523192 -0.11591841 -0.04964722 -0.13457249
#> Patient_10          NA          NA  0.33916636          NA  0.41290285
#> Patient_11  0.53468772  0.63695612  0.56587595  0.59103740  0.59656301
#> Patient_12 -0.22961660 -0.15914308 -0.19687857 -0.17901235          NA
#> Patient_13          NA  0.27092717  0.15318971  0.20044661          NA
#> Patient_14  0.23700457  0.33020565  0.26158199  0.29521261  0.36009615
#> Patient_15          NA  0.15166120  0.05639098  0.06982053  0.05015593
#> Patient_16 -0.43831037          NA          NA -0.31321620 -0.38520080
#> Patient_17 -0.69763097 -0.66074859 -0.72488519 -0.66989814 -0.71098493
#> Patient_18  0.67057692  0.72445889  0.68511570  0.67307692          NA
#> Patient_19 -0.08073696 -0.05541770          NA          NA -0.04266052
#> Patient_20 -0.50147456          NA -0.53063029 -0.49890631          NA
#> Patient_21  0.43000518  0.50327054  0.40301581  0.45101007  0.49784511
#> Patient_22 -0.23121960 -0.23320204          NA          NA -0.17612265
#> Patient_23  0.50207077  0.55696981  0.48411116  0.51573983  0.52137438
#> Patient_24          NA  0.43203536  0.28060885          NA  0.38894743
#> Patient_25 -0.88461538 -0.88624973          NA -0.88635376 -0.88461538
#> Patient_26          NA -0.54990122 -0.62511350 -0.47836126 -0.54253240
#> Patient_27  0.72557003          NA          NA  0.74331984  0.76374077
#> Patient_28 -0.32276383 -0.27737703          NA -0.27168576 -0.35125437
#> Patient_29  0.64749272  0.67972144          NA          NA  0.66119135
#> Patient_30  0.38449575          NA  0.29459135  0.38975372  0.40270798
#>                    [,6]        [,7]        [,8]        [,9]       [,10]
#> Patient_1   0.287892206  0.07731092          NA          NA          NA
#> Patient_2            NA  0.40490657  0.48065710          NA  0.40648911
#> Patient_3  -0.713863060          NA -0.74908112          NA -0.76551724
#> Patient_4   0.485377243  0.36076737  0.40403166  0.43316267  0.38833239
#> Patient_5  -0.641047688 -0.69069533          NA -0.67113559 -0.67383821
#> Patient_6   0.596352102  0.55937295          NA          NA          NA
#> Patient_7            NA -0.15551271 -0.11001488 -0.05044030 -0.14205794
#> Patient_8  -0.007484251 -0.05375705 -0.08480358 -0.09033141          NA
#> Patient_9   0.009386447 -0.11683653          NA -0.10876116 -0.04685369
#> Patient_10  0.444117474          NA  0.48044126  0.45232187  0.43006993
#> Patient_11  0.671872066          NA  0.56856187          NA  0.57246561
#> Patient_12 -0.164035811 -0.21286724          NA          NA -0.27964613
#> Patient_13  0.447729519  0.16545774  0.30348287  0.24917146  0.23532989
#> Patient_14           NA  0.27393162  0.39030673  0.34027149  0.34995403
#> Patient_15  0.157614345          NA  0.07809252  0.05874894  0.13523573
#> Patient_16 -0.340800652          NA -0.38362993          NA -0.34240740
#> Patient_17           NA -0.70018536 -0.67908543 -0.65975250 -0.66834456
#> Patient_18           NA  0.70197044  0.69161184  0.69070513  0.65384615
#> Patient_19  0.144051601          NA          NA -0.04086538 -0.08170015
#> Patient_20 -0.444967103 -0.46998124 -0.46255303 -0.49807692 -0.55373992
#> Patient_21  0.550699301  0.47548036  0.52666232          NA          NA
#> Patient_22 -0.162102448          NA -0.19325736 -0.22513642          NA
#> Patient_23  0.539898769  0.49604658          NA  0.53039852  0.49654755
#> Patient_24  0.501368315          NA  0.41337948  0.38616939  0.39907981
#> Patient_25 -0.884615385 -0.92307692 -0.90401115 -0.88461538          NA
#> Patient_26 -0.419018168 -0.49331104 -0.51731742 -0.50251123 -0.53099201
#> Patient_27           NA  0.75325733  0.77483444  0.78127589  0.74079157
#> Patient_28           NA -0.34474298 -0.24262487 -0.28445513          NA
#> Patient_29           NA  0.66354732          NA  0.69795371          NA
#> Patient_30  0.420801820  0.41047850  0.42000000  0.40045597  0.37525988
#>                   [,11]      [,12]       [,13]       [,14]      [,15]
#> Patient_1   0.231869929         NA  0.08741851  0.15280841  0.1346960
#> Patient_2            NA  0.5311133  0.36520147          NA  0.4230769
#> Patient_3            NA -0.7176773 -0.78352799          NA -0.7423840
#> Patient_4            NA         NA          NA  0.39058034  0.3253041
#> Patient_5  -0.675862299 -0.6619803          NA -0.67393603 -0.6315825
#> Patient_6   0.592925770  0.6137244  0.50073086          NA  0.5729996
#> Patient_7  -0.134196624         NA -0.23967052          NA         NA
#> Patient_8            NA  0.1109385 -0.31308298 -0.09801111 -0.2119439
#> Patient_9            NA  0.1153545 -0.14325988 -0.11383261 -0.1815562
#> Patient_10  0.467919447  0.5214042  0.38175156  0.43317009  0.3908571
#> Patient_11  0.594158953  0.5863409  0.55022328  0.61797824  0.5448563
#> Patient_12 -0.150349650 -0.1370079 -0.31563275          NA -0.2627612
#> Patient_13  0.261096909  0.4059829  0.15915426  0.24701386  0.2217230
#> Patient_14  0.283359498         NA  0.21336624  0.30696203  0.2985348
#> Patient_15           NA         NA  0.01330849  0.06908951         NA
#> Patient_16 -0.427350427 -0.2729917          NA          NA -0.4117941
#> Patient_17 -0.707819805 -0.6699993 -0.69578484 -0.69642350         NA
#> Patient_18  0.702676137         NA  0.66530278  0.70318475  0.7091503
#> Patient_19  0.006534583  0.1130536 -0.11365534 -0.14897574         NA
#> Patient_20 -0.474480784 -0.4216641 -0.53464771 -0.47212566 -0.4823574
#> Patient_21  0.471104704  0.6004121  0.42023164  0.49940727         NA
#> Patient_22 -0.218727534 -0.1713314 -0.28200972 -0.27229089         NA
#> Patient_23           NA  0.5311741  0.46402545  0.49774115  0.5206668
#> Patient_24           NA         NA          NA  0.35196501  0.3823305
#> Patient_25 -0.884615385 -0.9041554          NA -0.88461538 -0.8863883
#> Patient_26 -0.512820513 -0.3796044          NA -0.52862633         NA
#> Patient_27  0.769550749  0.7733783  0.72691992  0.77395577         NA
#> Patient_28 -0.292259084 -0.1972809 -0.37053796          NA -0.3380893
#> Patient_29  0.673375196  0.7283994          NA          NA  0.6633358
#> Patient_30  0.405136205         NA          NA  0.38105688  0.3485086
#>                 [,16]       [,17]       [,18]       [,19]      [,20]
#> Patient_1   0.1451049  0.13731615  0.19619143  0.06526012         NA
#> Patient_2          NA  0.49309006  0.48826155  0.39880342  0.3148460
#> Patient_3          NA -0.75101215 -0.72939560 -0.76048951 -0.7661996
#> Patient_4   0.3801619          NA  0.34733032          NA  0.2574957
#> Patient_5  -0.7021816 -0.64988105 -0.63033081 -0.68703259         NA
#> Patient_6   0.5634209          NA  0.53888221  0.50152383         NA
#> Patient_7  -0.1738877 -0.10563568 -0.13621660 -0.24413839         NA
#> Patient_8          NA -0.06129426 -0.08814715 -0.23117148 -0.3446617
#> Patient_9  -0.1272562  0.03116401 -0.10162368 -0.16172063 -0.1975277
#> Patient_10         NA          NA          NA          NA  0.3662154
#> Patient_11  0.5787061  0.58282559  0.59073783  0.55533499  0.5204583
#> Patient_12 -0.2117465          NA          NA -0.26866792 -0.2996537
#> Patient_13  0.1735693  0.34527439  0.30276877          NA  0.0937817
#> Patient_14  0.2648141  0.38346366  0.27689911  0.24774317  0.3064655
#> Patient_15         NA  0.13141575          NA          NA  0.0184988
#> Patient_16         NA -0.31949501 -0.34766214          NA         NA
#> Patient_17 -0.6767399 -0.62315392 -0.68846506          NA -0.7502870
#> Patient_18  0.6858183          NA  0.73852459  0.69816731  0.6425658
#> Patient_19 -0.1151793 -0.04698512          NA -0.15455951         NA
#> Patient_20 -0.4867054 -0.44372470          NA -0.49163880 -0.5337188
#> Patient_21         NA  0.50975691  0.49650644  0.47094017  0.4149903
#> Patient_22         NA -0.17347692          NA -0.29683775 -0.3073844
#> Patient_23  0.5027055  0.50281134  0.52250574  0.44205679  0.4339421
#> Patient_24  0.4209768  0.41796899          NA          NA  0.2360346
#> Patient_25 -0.9041912          NA -0.88461538          NA         NA
#> Patient_26 -0.5608866 -0.51676195 -0.48259452 -0.59216255 -0.6097187
#> Patient_27  0.7459016  0.79258010          NA  0.72784730  0.7189542
#> Patient_28 -0.3758154          NA -0.35358217 -0.34935897 -0.3145131
#> Patient_29  0.6460553          NA  0.69186633  0.64479451  0.6424605
#> Patient_30  0.3671220  0.38501379  0.38615385  0.31889764         NA
#> 
#> $PERMUTATIONS$`2018_Tiriac_PDAC_PDO_basal-like_signature`
#>                  [,1]        [,2]        [,3]        [,4]        [,5]
#> Patient_1  -0.1293820 -0.30238095 -0.15214180          NA -0.18025890
#> Patient_2  -0.3385965 -0.40488798 -0.42979798 -0.44887064          NA
#> Patient_3          NA  0.82050147  0.81029412          NA          NA
#> Patient_4  -0.3634228          NA -0.38048246 -0.42428811 -0.42827529
#> Patient_5   0.8301724  0.75498354          NA  0.82391304  0.82549445
#> Patient_6  -0.5882960 -0.61338028 -0.56673004 -0.63509532          NA
#> Patient_7   0.1619492          NA  0.20888562  0.11439280          NA
#> Patient_8          NA          NA  0.33407960          NA  0.25494505
#> Patient_9          NA  0.10311284  0.19166667  0.09212598  0.17301587
#> Patient_10         NA          NA -0.44405374          NA -0.43983364
#> Patient_11 -0.6387114 -0.70349099 -0.65806452 -0.67564103 -0.62583732
#> Patient_12  0.2550201  0.11180812  0.29210526  0.12652200          NA
#> Patient_13         NA -0.30636704 -0.11964286 -0.30185185          NA
#> Patient_14 -0.3194030 -0.36001178 -0.36013072 -0.38596491 -0.36839484
#> Patient_15         NA -0.07586912 -0.09785276 -0.05314465 -0.01441103
#> Patient_16  0.5363118          NA          NA  0.43148148  0.44641745
#> Patient_17  0.7458663  0.73463542  0.77734878  0.76374889  0.75000000
#> Patient_18 -0.7618901 -0.81825397 -0.78478729 -0.79657534          NA
#> Patient_19  0.2470880  0.12215113          NA          NA  0.18364780
#> Patient_20  0.6022613          NA  0.63436344  0.57601487          NA
#> Patient_21 -0.5400656 -0.60386330 -0.54614086 -0.57352941 -0.62307501
#> Patient_22  0.3065728  0.27520793          NA          NA  0.32927429
#> Patient_23 -0.4850967 -0.61410256 -0.52607076 -0.54794829 -0.61406423
#> Patient_24         NA -0.40249554 -0.29215686          NA -0.34813084
#> Patient_25  0.9351970  0.93164557          NA  0.93519695  0.94967742
#> Patient_26         NA  0.60202492  0.70980392  0.54433870  0.61417166
#> Patient_27 -0.7792271          NA          NA -0.77269273 -0.78174603
#> Patient_28  0.4287472  0.35915033          NA  0.37208931  0.36159420
#> Patient_29 -0.7099781 -0.77209302          NA          NA -0.72598684
#> Patient_30 -0.3653641          NA -0.23137083 -0.46289954 -0.47201946
#>                   [,6]        [,7]        [,8]        [,9]       [,10]
#> Patient_1  -0.29225908 -0.08968134          NA          NA          NA
#> Patient_2           NA -0.38708087 -0.44916143          NA -0.41829574
#> Patient_3   0.77927928          NA  0.80484983          NA  0.82677794
#> Patient_4  -0.56876047 -0.42124183 -0.49256198 -0.47641196 -0.48641189
#> Patient_5   0.77546161  0.83712121          NA  0.82681159  0.83515284
#> Patient_6  -0.64455782 -0.62997602          NA          NA          NA
#> Patient_7           NA  0.11845238  0.10454545  0.12080182  0.18756004
#> Patient_8   0.09749035  0.18820698  0.15914972  0.16449772          NA
#> Patient_9   0.02458746  0.16004016          NA  0.10618280 -0.00942029
#> Patient_10 -0.49741697          NA -0.53401908 -0.53168800 -0.51150729
#> Patient_11 -0.72473118          NA -0.61587545          NA -0.63996360
#> Patient_12  0.10497238  0.20651629          NA          NA  0.20152488
#> Patient_13 -0.38523447 -0.17921348 -0.32766990 -0.28269962 -0.30662100
#> Patient_14          NA -0.36213922 -0.40516224 -0.35191638 -0.46936027
#> Patient_15 -0.17343358          NA -0.05576441  0.01081081 -0.10663189
#> Patient_16  0.41793169          NA  0.41992337          NA  0.39687500
#> Patient_17          NA  0.75194805  0.75785340  0.72825607  0.77129145
#> Patient_18          NA -0.79387216 -0.81240126 -0.80852594 -0.82923417
#> Patient_19 -0.08680927          NA          NA  0.12439479  0.21824058
#> Patient_20  0.55222965  0.56941303  0.57447046  0.57011494  0.66974110
#> Patient_21 -0.67299546 -0.58484848 -0.62750374          NA          NA
#> Patient_22  0.20931559          NA  0.26199518  0.23277571          NA
#> Patient_23 -0.66391586 -0.50168350          NA -0.57339956 -0.53506045
#> Patient_24 -0.54256595          NA -0.41369048 -0.41880190 -0.37543860
#> Patient_25  0.94601542  0.96060991  0.92708333  0.94601542          NA
#> Patient_26  0.51399491  0.64721030  0.60980392  0.59817518  0.62740741
#> Patient_27          NA -0.79049919 -0.80327613 -0.81059190 -0.78778468
#> Patient_28          NA  0.40476190  0.29000000  0.38247366          NA
#> Patient_29          NA -0.74805195          NA -0.74199134          NA
#> Patient_30 -0.44637681 -0.39905009 -0.48105647 -0.45303030 -0.40126904
#>                 [,11]       [,12]      [,13]       [,14]      [,15]      [,16]
#> Patient_1  -0.1640575          NA -0.1517937 -0.20965309 -0.1797764 -0.1496974
#> Patient_2          NA -0.50440252 -0.4101928          NA -0.3962199         NA
#> Patient_3          NA  0.78602794  0.8414651          NA  0.8028986         NA
#> Patient_4          NA          NA         NA -0.36808511 -0.3860438 -0.4801846
#> Patient_5   0.8376437  0.80619048         NA  0.82676399  0.7725420  0.8482759
#> Patient_6  -0.6199324 -0.65289855 -0.5253697          NA -0.5878378 -0.6201754
#> Patient_7   0.1255725          NA  0.2336469          NA         NA  0.1158371
#> Patient_8          NA  0.01598407  0.3448187  0.18333333  0.2728727         NA
#> Patient_9          NA -0.04673660  0.1845052  0.08148148  0.2271150  0.2126506
#> Patient_10 -0.5214953 -0.57385159 -0.4750000 -0.44304813 -0.4070142         NA
#> Patient_11 -0.6158385 -0.67046414 -0.6297657 -0.65655608 -0.6662252 -0.6549922
#> Patient_12  0.2327402  0.04067358  0.2926929          NA  0.2249548  0.1788288
#> Patient_13 -0.2461905 -0.35726341 -0.1398551 -0.19064171 -0.2580756 -0.2279762
#> Patient_14 -0.3563768          NA -0.2775081 -0.31495726 -0.3386814 -0.3303922
#> Patient_15         NA          NA  0.0463925 -0.01597542         NA         NA
#> Patient_16  0.4607672  0.34259259         NA          NA  0.4907407         NA
#> Patient_17  0.7408694  0.73072917  0.7748366  0.73888889         NA  0.7468736
#> Patient_18 -0.8128931          NA -0.7612440 -0.78467310 -0.7838667 -0.7929548
#> Patient_19  0.1410714 -0.07857143  0.2892367  0.20274725         NA  0.2057621
#> Patient_20  0.5835609  0.54134615  0.6921598  0.59926160  0.5605964  0.5910468
#> Patient_21 -0.6166667 -0.72050754 -0.5885714 -0.60785609         NA         NA
#> Patient_22  0.2671091  0.15654762  0.3052846  0.31358025         NA         NA
#> Patient_23         NA -0.56623236 -0.4357418 -0.54052066 -0.5876823 -0.5104794
#> Patient_24         NA          NA         NA -0.37271386 -0.3595771 -0.3674603
#> Patient_25  0.9269521  0.93661616         NA  0.94358974  0.9363868  0.9390295
#> Patient_26  0.6379643  0.52101449         NA  0.62415170         NA  0.6721973
#> Patient_27 -0.8180835 -0.83499496 -0.7880952 -0.79978769         NA -0.7743508
#> Patient_28  0.3971698  0.24974671  0.4798667          NA  0.3768473  0.4242905
#> Patient_29 -0.7730769 -0.78333333         NA          NA -0.7562667 -0.7300000
#> Patient_30 -0.3813933          NA         NA -0.39818841 -0.3821739 -0.3943953
#>                  [,17]       [,18]      [,19]       [,20]
#> Patient_1  -0.19420290 -0.21578100 -0.1411111          NA
#> Patient_2  -0.41759628 -0.46015326 -0.4070140 -0.34797768
#> Patient_3   0.81376812  0.78546310  0.8280142  0.84686469
#> Patient_4           NA -0.43635308         NA -0.32981575
#> Patient_5   0.83199513  0.77913652  0.8549737          NA
#> Patient_6           NA -0.61376684 -0.5132027          NA
#> Patient_7   0.15748446  0.06666667  0.1865530          NA
#> Patient_8   0.18627617  0.15555556  0.2975248  0.36012514
#> Patient_9  -0.06243339  0.12183236  0.1722222  0.19133858
#> Patient_10          NA          NA         NA -0.46597106
#> Patient_11 -0.64314495 -0.66923077 -0.6412151 -0.59425837
#> Patient_12          NA          NA  0.2554927  0.26005510
#> Patient_13 -0.35651769 -0.26327575         NA -0.09464286
#> Patient_14 -0.45632184 -0.35943953 -0.2697966 -0.39257703
#> Patient_15 -0.14831625          NA         NA  0.03225108
#> Patient_16  0.40312307  0.45691523         NA          NA
#> Patient_17  0.74600000  0.74190853         NA  0.77734878
#> Patient_18          NA -0.84353471 -0.7813008 -0.79890966
#> Patient_19  0.14009324          NA  0.2945946          NA
#> Patient_20  0.60130293          NA  0.6380296  0.69812030
#> Patient_21 -0.69414025 -0.58484848 -0.6119784 -0.52488061
#> Patient_22  0.17584864          NA  0.2842429  0.37666667
#> Patient_23 -0.54490823 -0.56038251 -0.3994505 -0.43127413
#> Patient_24 -0.43527397          NA         NA -0.30991926
#> Patient_25          NA  0.93046776         NA          NA
#> Patient_26  0.62744361  0.54312407  0.7234017  0.77134293
#> Patient_27 -0.83521870          NA -0.8122257 -0.77680000
#> Patient_28          NA  0.31644295  0.4629333  0.40707871
#> Patient_29          NA -0.77445681 -0.7110763 -0.68354978
#> Patient_30 -0.34915825 -0.44581056 -0.3104491          NA
#> 
#> 
#> $SD
#> $SD$`2018_Tiriac_PDAC_PDO_classical_signature`
#>  Patient_1  Patient_2  Patient_3  Patient_4  Patient_5  Patient_6  Patient_7 
#> 0.06761385 0.06714942 0.02178279 0.05565734 0.02417271 0.03563254 0.05831655 
#>  Patient_8  Patient_9 Patient_10 Patient_11 Patient_12 Patient_13 Patient_14 
#> 0.12562910 0.08134402 0.05168530 0.03644722 0.05833163 0.09260057 0.04986102 
#> Patient_15 Patient_16 Patient_17 Patient_18 Patient_19 Patient_20 Patient_21 
#> 0.04966692 0.05169848 0.02905760 0.02493697 0.09060446 0.03577008 0.05241450 
#> Patient_22 Patient_23 Patient_24 Patient_25 Patient_26 Patient_27 Patient_28 
#> 0.05036711 0.03207177 0.06963245 0.01203991 0.06210010 0.02327648 0.05069715 
#> Patient_29 Patient_30 
#> 0.02596018 0.03477593 
#> 
#> $SD$`2018_Tiriac_PDAC_PDO_basal-like_signature`
#>  Patient_1  Patient_2  Patient_3  Patient_4  Patient_5  Patient_6  Patient_7 
#> 0.05873195 0.04316848 0.02178375 0.06434722 0.02987455 0.04337084 0.04754415 
#>  Patient_8  Patient_9 Patient_10 Patient_11 Patient_12 Patient_13 Patient_14 
#> 0.09961950 0.09029813 0.04920713 0.03166041 0.07673620 0.08636134 0.05151586 
#> Patient_15 Patient_16 Patient_17 Patient_18 Patient_19 Patient_20 Patient_21 
#> 0.06945595 0.05114272 0.01614104 0.02240488 0.11941258 0.04751882 0.05354357 
#> Patient_22 Patient_23 Patient_24 Patient_25 Patient_26 Patient_27 Patient_28 
#> 0.06222335 0.06797764 0.06495541 0.00942914 0.07028981 0.02082624 0.06139822 
#> Patient_29 Patient_30 
#> 0.03114349 0.06515294 
#> 
#>