TODO

TODO.

runSubtyping(geneLists, expectedCountsMatrix, permRatio = 0.75, permNbr = 20)

Arguments

geneLists

a list containing signature gene lists. At least two signature gene lists must be present.

expectedCountsMatrix

a normalized gene expression data matrix with rows corresponding to genes and columns to samples.

permRatio

a numeric between 0 and 1 representing the number of samples that are retained for the permutation step. The rounded value of the number of patients multiplied by this numeric should be inferior to the total number of patients so that the permutation step is not done on the entire cohort. Default: 0.75.

permNbr

a integer, 5 or higher, representing the number of permutation sampling done. In addition, the number of permutations must be equal or inferior to the total number of unique permutations with the dataset considering the parameters selected by the user. Default: 20.

Value

TRUE

Author

Astrid Deschênes

Examples

## Loading signatures
data("signatures")

## Load demo normalized expected counts for 30 patients
data("expNormalCountsDemo")

## Some of the enrichment results present in the dataset
## TODO
runSubtyping(geneLists=signatures, expectedCountsMatrix=expNormalCountsDemo, 
    permRatio=0.75, permNbr=20)
#> $GSVA_RESULTS
#>                                            Patient_1  Patient_2  Patient_3
#> 2018_Tiriac_PDAC_PDO_classical_signature   0.1422663  0.4274225 -0.7532768
#> 2018_Tiriac_PDAC_PDO_basal-like_signature -0.2385417 -0.4403292  0.8006550
#>                                            Patient_4  Patient_5  Patient_6
#> 2018_Tiriac_PDAC_PDO_classical_signature   0.3479971 -0.6732937  0.5735090
#> 2018_Tiriac_PDAC_PDO_basal-like_signature -0.4500838  0.8175287 -0.6076389
#>                                            Patient_7  Patient_8  Patient_9
#> 2018_Tiriac_PDAC_PDO_classical_signature  -0.1247647 -0.1261704 -0.1287452
#> 2018_Tiriac_PDAC_PDO_basal-like_signature  0.1209839  0.2276770  0.1643817
#>                                           Patient_10 Patient_11 Patient_12
#> 2018_Tiriac_PDAC_PDO_classical_signature   0.4370278  0.5913501 -0.2002262
#> 2018_Tiriac_PDAC_PDO_basal-like_signature -0.4448441 -0.6485138  0.1993417
#>                                           Patient_13 Patient_14   Patient_15
#> 2018_Tiriac_PDAC_PDO_classical_signature   0.2448591  0.3249211  0.053730189
#> 2018_Tiriac_PDAC_PDO_basal-like_signature -0.2308458 -0.3733219 -0.005216285
#>                                           Patient_16 Patient_17 Patient_18
#> 2018_Tiriac_PDAC_PDO_classical_signature  -0.3930325 -0.7043936  0.7082490
#> 2018_Tiriac_PDAC_PDO_basal-like_signature  0.4381176  0.7356150 -0.8162717
#>                                            Patient_19 Patient_20 Patient_21
#> 2018_Tiriac_PDAC_PDO_classical_signature  -0.05232661 -0.4716993  0.4735043
#> 2018_Tiriac_PDAC_PDO_basal-like_signature  0.14282238  0.5886179 -0.6005646
#>                                           Patient_22 Patient_23 Patient_24
#> 2018_Tiriac_PDAC_PDO_classical_signature  -0.2313920  0.5079302  0.4728932
#> 2018_Tiriac_PDAC_PDO_basal-like_signature  0.2876812 -0.5792517 -0.4255245
#>                                           Patient_25 Patient_26 Patient_27
#> 2018_Tiriac_PDAC_PDO_classical_signature  -0.8846154 -0.5213023  0.7477403
#> 2018_Tiriac_PDAC_PDO_basal-like_signature  0.9387755  0.6288557 -0.7916667
#>                                           Patient_28 Patient_29 Patient_30
#> 2018_Tiriac_PDAC_PDO_classical_signature  -0.2808672  0.6611142  0.3823810
#> 2018_Tiriac_PDAC_PDO_basal-like_signature  0.3847667 -0.7322511 -0.4204301
#> attr(,"gsvaParam")
#> attr(,"gsvaParam")$originalClassWasSE
#> [1] FALSE
#> 
#> attr(,"gsvaParam")$geneSets
#> attr(,"gsvaParam")$geneSets$`2018_Tiriac_PDAC_PDO_classical_signature`
#>  [1] "MUC13"      "GPR35"      "USH1C"      "BAIAP2L2"   "GAL3ST1"   
#>  [6] "VIL1"       "DDC"        "LGALS4"     "TSPAN8"     "SLC5A1"    
#> [11] "SEMA4G"     "TNFRSF11A"  "HSD17B2"    "CKMT1A"     "DPCR1"     
#> [16] "GCNT3"      "SLC6A20"    "POF1B"      "NR1I2"      "MUC17"     
#> [21] "HHLA2"      "BTNL8"      "HEPH"       "PDZD3"      "MYO1A"     
#> [26] "CDH17"      "MYO7B"      "GPA33"      "CALML4"     "ANXA13"    
#> [31] "TM4SF20"    "SI"         "SLC44A4"    "FAM83E"     "SLC40A1"   
#> [36] "CTSE"       "BACE2"      "SULT1C2"    "AGR3"       "ATP7B"     
#> [41] "LYZ"        "TESC"       "TOX3"       "SH3BGRL2"   "GMDS"      
#> [46] "GATA6"      "GPX2"       "CA2"        "ST6GALNAC1" "FOXA3"     
#> [51] "AKR1C3"     "TMC5"       "CYP2S1"     "PAQR8"      "HKDC1"     
#> [56] "IYD"        "DOK4"       "CYP2C18"    "KCNE3"      "FZD5"      
#> [61] "RASEF"      "GKN1"      
#> 
#> attr(,"gsvaParam")$geneSets$`2018_Tiriac_PDAC_PDO_basal-like_signature`
#>  [1] "ANO1"     "ADORA2B"  "S100A2"   "CSF2"     "ANXA1"    "EREG"    
#>  [7] "RGS20"    "CELSR1"   "WNT7B"    "AMIGO2"   "PTHLH"    "C12orf54"
#> [13] "TGM2"     "REEP3"    "DCBLD2"   "NMB"      "IKZF2"    "GPR87"   
#> [19] "MUC16"    "IVL"      "SPRR1B"   "BCAR3"    "TTC9"     "PTGES"   
#> [25] "CST6"     "CATSPER1"
#> 
#> 
#> attr(,"gsvaParam")$assay
#> [1] NA
#> 
#> attr(,"gsvaParam")$annotation
#> geneIdType: Null
#> 
#> attr(,"gsvaParam")$minSize
#> [1] 1
#> 
#> attr(,"gsvaParam")$maxSize
#> [1] Inf
#> 
#> attr(,"gsvaParam")$nzcount
#> [1] 2580
#> 
#> attr(,"gsvaParam")$ondisk
#> [1] "auto"
#> 
#> attr(,"gsvaParam")$checkNA
#> [1] "auto"
#> 
#> attr(,"gsvaParam")$didCheckNA
#> [1] TRUE
#> 
#> attr(,"gsvaParam")$anyNA
#> [1] FALSE
#> 
#> attr(,"gsvaParam")$use
#> [1] "everything"
#> 
#> attr(,"gsvaParam")$kcdf
#> [1] "auto"
#> 
#> attr(,"gsvaParam")$kcdfNoneMinSampleSize
#> [1] 200
#> 
#> attr(,"gsvaParam")$tau
#> [1] 1
#> 
#> attr(,"gsvaParam")$maxDiff
#> [1] TRUE
#> 
#> attr(,"gsvaParam")$absRanking
#> [1] FALSE
#> 
#> attr(,"gsvaParam")$sparse
#> [1] FALSE
#> 
#> attr(,"gsvaParam")$filterRows
#> [1] TRUE
#> 
#> attr(,"assay")
#> [1] "es"
#> attr(,"geneSets")
#> attr(,"geneSets")$`2018_Tiriac_PDAC_PDO_classical_signature`
#>  [1] "MUC13"      "GPR35"      "USH1C"      "BAIAP2L2"   "GAL3ST1"   
#>  [6] "VIL1"       "DDC"        "LGALS4"     "TSPAN8"     "SLC5A1"    
#> [11] "SEMA4G"     "TNFRSF11A"  "HSD17B2"    "CKMT1A"     "GCNT3"     
#> [16] "SLC6A20"    "POF1B"      "NR1I2"      "MUC17"      "HHLA2"     
#> [21] "BTNL8"      "HEPH"       "MYO1A"      "CDH17"      "MYO7B"     
#> [26] "GPA33"      "CALML4"     "ANXA13"     "TM4SF20"    "SI"        
#> [31] "SLC44A4"    "FAM83E"     "SLC40A1"    "CTSE"       "BACE2"     
#> [36] "SULT1C2"    "AGR3"       "ATP7B"      "LYZ"        "TESC"      
#> [41] "TOX3"       "SH3BGRL2"   "GMDS"       "GATA6"      "GPX2"      
#> [46] "CA2"        "ST6GALNAC1" "FOXA3"      "AKR1C3"     "TMC5"      
#> [51] "CYP2S1"     "PAQR8"      "HKDC1"      "IYD"        "DOK4"      
#> [56] "CYP2C18"    "KCNE3"      "FZD5"       "RASEF"      "GKN1"      
#> 
#> attr(,"geneSets")$`2018_Tiriac_PDAC_PDO_basal-like_signature`
#>  [1] "ANO1"     "ADORA2B"  "S100A2"   "CSF2"     "ANXA1"    "EREG"    
#>  [7] "RGS20"    "CELSR1"   "WNT7B"    "AMIGO2"   "PTHLH"    "C12orf54"
#> [13] "TGM2"     "REEP3"    "DCBLD2"   "NMB"      "IKZF2"    "GPR87"   
#> [19] "MUC16"    "IVL"      "SPRR1B"   "BCAR3"    "TTC9"     "PTGES"   
#> [25] "CST6"     "CATSPER1"
#> 
#> 
#> $BOOTSTRAP
#> $BOOTSTRAP$`2018_Tiriac_PDAC_PDO_classical_signature`
#>                   [,1]        [,2]        [,3]         [,4]          [,5]
#> Patient_1   0.25531915  0.15629314          NA           NA  0.1940098200
#> Patient_2   0.46821036  0.39947947  0.47575859  0.352554662  0.5087807889
#> Patient_3  -0.71069387 -0.75041770          NA           NA -0.7138630600
#> Patient_4           NA  0.32344083  0.30491673  0.227163462  0.3917577903
#> Patient_5  -0.62049072 -0.66175288 -0.67892108 -0.707228117 -0.6571011027
#> Patient_6   0.58689680          NA  0.53582964  0.495516753            NA
#> Patient_7           NA -0.11066359          NA -0.270757284            NA
#> Patient_8  -0.09653932 -0.20115115 -0.09284028 -0.332944833  0.0002327476
#> Patient_9  -0.10115215 -0.08717949          NA -0.188717949 -0.0149005657
#> Patient_10  0.42088420  0.46130086  0.44462713  0.286123680  0.4719788792
#> Patient_11  0.60317362          NA  0.55545617  0.552469894  0.6303963087
#> Patient_12 -0.23135110          NA          NA -0.252388535            NA
#> Patient_13          NA  0.23726572  0.20673442  0.078928740  0.3648368568
#> Patient_14  0.34832409  0.36920660  0.34226884  0.252826247  0.3918174666
#> Patient_15  0.11735537          NA  0.02441696 -0.002192275            NA
#> Patient_16 -0.29621924 -0.34589417 -0.41310105           NA            NA
#> Patient_17 -0.67941576 -0.67052827 -0.70035613 -0.708416123 -0.6739926740
#> Patient_18  0.70142285          NA  0.67307692  0.634265734            NA
#> Patient_19          NA -0.06437806 -0.06268029 -0.271926671  0.0268874644
#> Patient_20          NA          NA -0.48998690           NA -0.4363177806
#> Patient_21  0.52024626  0.55824925  0.46087590  0.395129562  0.5120026525
#> Patient_22 -0.21931637 -0.22302738 -0.20501880 -0.291633466 -0.2247454973
#> Patient_23  0.56017370          NA          NA  0.360051713  0.5155021969
#> Patient_24  0.44429708  0.41673581          NA  0.167987322  0.4427193602
#> Patient_25 -0.86508050          NA -0.88461538           NA -0.8846153846
#> Patient_26          NA -0.54020979 -0.52083196 -0.600160256 -0.4748210716
#> Patient_27          NA  0.78529657  0.76916803           NA  0.7885906040
#> Patient_28 -0.27531365 -0.24880182 -0.24281401           NA            NA
#> Patient_29          NA  0.69135724  0.67490400  0.647996017            NA
#> Patient_30  0.40333478  0.37968598          NA           NA  0.4254280203
#>                   [,6]         [,7]        [,8]         [,9]        [,10]
#> Patient_1   0.06028414  0.147307692  0.13027295  0.200851172           NA
#> Patient_2           NA  0.456305405  0.42756794  0.471153846  0.451285828
#> Patient_3  -0.79710593 -0.748502994 -0.74583527 -0.742154368           NA
#> Patient_4   0.23360483  0.424639902  0.30832542  0.421874036  0.351648352
#> Patient_5  -0.67526606           NA -0.66127435           NA -0.702344066
#> Patient_6           NA  0.584380222          NA  0.563327089           NA
#> Patient_7           NA -0.072712884 -0.15579399 -0.088852673           NA
#> Patient_8  -0.31093906           NA          NA  0.008579433 -0.203870117
#> Patient_9  -0.20342008           NA -0.08910891 -0.024611922 -0.168699876
#> Patient_10          NA           NA  0.46810433  0.446056321  0.374193166
#> Patient_11  0.54365054           NA  0.52065527           NA  0.556931769
#> Patient_12          NA -0.197849527          NA -0.118659205 -0.281430288
#> Patient_13  0.15692071  0.198431212  0.25478950  0.320241076  0.123826672
#> Patient_14  0.25136038  0.316596587  0.35308551  0.411674208  0.281195238
#> Patient_15  0.04890206  0.106788079          NA  0.098984279  0.005166475
#> Patient_16 -0.46823515 -0.392388300 -0.38954484 -0.326704048 -0.395232524
#> Patient_17          NA -0.708398024 -0.64446013 -0.702706553 -0.701869559
#> Patient_18  0.67987065           NA  0.67052980  0.721273341  0.640080301
#> Patient_19 -0.26394598 -0.006410256          NA           NA           NA
#> Patient_20          NA -0.466010733 -0.48250328 -0.445767361           NA
#> Patient_21  0.46436896           NA  0.50282467           NA           NA
#> Patient_22 -0.24212454 -0.183316683 -0.20261377 -0.234884823 -0.271771240
#> Patient_23  0.37942589  0.540418246          NA           NA  0.502354788
#> Patient_24  0.21523959  0.444454179  0.39560440           NA  0.326151517
#> Patient_25          NA -0.886180045          NA -0.923076923 -0.884615385
#> Patient_26 -0.60807658           NA -0.54388904 -0.406421405           NA
#> Patient_27  0.75448613  0.772614108  0.74153846  0.779983457  0.738037307
#> Patient_28 -0.35922270 -0.224291498 -0.31243914           NA -0.371333702
#> Patient_29  0.64243176  0.667753151          NA  0.702487778  0.680190931
#> Patient_30  0.22805643  0.438653247  0.38224572           NA  0.363558064
#>                  [,11]       [,12]        [,13]        [,14]       [,15]
#> Patient_1   0.39874662  0.19567753  0.264205802  0.207743335  0.15899195
#> Patient_2           NA  0.46114253           NA  0.439954325  0.39003408
#> Patient_3  -0.71112110 -0.73195021 -0.710277957           NA -0.73950748
#> Patient_4   0.47272617  0.48446626  0.468429875  0.418135270  0.38340081
#> Patient_5  -0.57323814 -0.64548779 -0.628205128 -0.689429618          NA
#> Patient_6           NA  0.59753175  0.605984271           NA  0.56132577
#> Patient_7   0.04793099  0.04598232  0.005434431  0.031126258          NA
#> Patient_8   0.05942985  0.02749765 -0.024004789 -0.041731782 -0.15384615
#> Patient_9   0.04941777          NA  0.033360083 -0.031127409          NA
#> Patient_10  0.51208124  0.45963502  0.496230416  0.428193366  0.34554720
#> Patient_11  0.63148920  0.61722226  0.616548158  0.655365622  0.55208465
#> Patient_12          NA -0.11452729 -0.205148253           NA -0.24164724
#> Patient_13  0.41631633          NA  0.339874411  0.364033575  0.19941875
#> Patient_14          NA  0.37436302  0.354116707           NA  0.33860759
#> Patient_15          NA  0.16937370           NA  0.143141026  0.04527144
#> Patient_16 -0.30780661 -0.26523171 -0.318774446 -0.327600115 -0.43732843
#> Patient_17 -0.70669568 -0.74840040 -0.685346646 -0.728402910 -0.71529827
#> Patient_18  0.78062914          NA           NA  0.740345111  0.67867656
#> Patient_19  0.08400277          NA  0.097882961  0.002125429          NA
#> Patient_20 -0.42041397 -0.41606700 -0.426489578 -0.436202686 -0.48583328
#> Patient_21  0.63432053  0.47648466           NA  0.511929596  0.41181507
#> Patient_22 -0.05150326          NA           NA           NA          NA
#> Patient_23          NA  0.53339587  0.537451055  0.575612952  0.50506235
#> Patient_24  0.53296703  0.48379601           NA  0.479085431          NA
#> Patient_25 -0.88461538 -0.92307692 -0.884615385 -0.884615385 -0.92307692
#> Patient_26 -0.40041266 -0.45503514           NA -0.451019618 -0.53836433
#> Patient_27          NA  0.76891106  0.779275484           NA  0.73491028
#> Patient_28 -0.15714638          NA -0.215906893 -0.284203407          NA
#> Patient_29  0.73875789          NA           NA           NA          NA
#> Patient_30          NA          NA  0.434284664           NA  0.35796036
#>                  [,16]        [,17]       [,18]       [,19]       [,20]
#> Patient_1   0.24027517  0.232110550          NA  0.10464081  0.11950437
#> Patient_2           NA  0.485140495          NA  0.32568962  0.32595462
#> Patient_3  -0.73970137           NA -0.72894084 -0.76084642 -0.78234266
#> Patient_4   0.32094532  0.403634161  0.45063598  0.25095295          NA
#> Patient_5  -0.65553306 -0.648313788 -0.62758299          NA          NA
#> Patient_6   0.52836538           NA  0.60904044  0.45551363  0.56374971
#> Patient_7           NA           NA          NA -0.24302043          NA
#> Patient_8  -0.11931171  0.039976690 -0.01313829 -0.24731280 -0.24529277
#> Patient_9  -0.04735522 -0.020275613          NA -0.12213326          NA
#> Patient_10  0.41819481  0.489755298          NA          NA  0.35891595
#> Patient_11          NA           NA          NA  0.55069212  0.54184809
#> Patient_12          NA           NA -0.11568254 -0.26855493 -0.32009926
#> Patient_13  0.32047382  0.339492586  0.37009063          NA  0.17518707
#> Patient_14  0.31855274  0.423251484  0.35998807  0.28687196  0.22284891
#> Patient_15  0.12515522  0.255167183  0.15891766  0.04797847  0.01587051
#> Patient_16 -0.36229238 -0.339226711 -0.35828136 -0.37197376 -0.47638995
#> Patient_17 -0.63621048 -0.677591355 -0.67724211 -0.70693788          NA
#> Patient_18  0.70719603  0.754560531  0.73198921  0.66368368  0.68729904
#> Patient_19 -0.06049144 -0.006405448  0.09397426          NA -0.21297750
#> Patient_20 -0.48081687 -0.427115695 -0.42565124          NA          NA
#> Patient_21  0.55144352           NA  0.53143303  0.45203620  0.41196726
#> Patient_22 -0.15712464 -0.151776475 -0.10226684 -0.25227427          NA
#> Patient_23          NA           NA  0.59707855  0.47086954  0.51308612
#> Patient_24  0.45351171  0.489044113  0.50271950          NA  0.28677264
#> Patient_25          NA -0.884615385 -0.88461538          NA -0.88461538
#> Patient_26 -0.48974806 -0.449372703          NA -0.59706174 -0.62240691
#> Patient_27          NA           NA          NA  0.72735605          NA
#> Patient_28          NA -0.231444533 -0.24057863 -0.32371019 -0.39927885
#> Patient_29  0.69321115  0.737919738  0.71936383          NA  0.65042735
#> Patient_30  0.35031348  0.456615859  0.42696495  0.33307024  0.30263158
#> 
#> $BOOTSTRAP$`2018_Tiriac_PDAC_PDO_basal-like_signature`
#>                   [,1]        [,2]        [,3]       [,4]         [,5]
#> Patient_1  -0.26562998 -0.16966292          NA         NA -0.260682493
#> Patient_2  -0.44675052 -0.45298651 -0.49614035 -0.3906314 -0.505230126
#> Patient_3   0.81428571  0.81652340          NA         NA  0.809459459
#> Patient_4           NA -0.46613162 -0.41156463 -0.2857938 -0.492616034
#> Patient_5   0.77157716  0.80487399  0.82588857  0.8368408  0.800144718
#> Patient_6  -0.64723708          NA -0.59123742 -0.5472477           NA
#> Patient_7           NA  0.16257310          NA  0.2797101           NA
#> Patient_8   0.18984064  0.29045936  0.18039216  0.3205219  0.111638204
#> Patient_9   0.11136208  0.03026961          NA  0.2788577  0.009119497
#> Patient_10 -0.48239564 -0.50511432 -0.46397650 -0.4359466 -0.498589065
#> Patient_11 -0.66414474          NA -0.59031008 -0.6240369 -0.663291139
#> Patient_12  0.18858318          NA          NA  0.3755481           NA
#> Patient_13          NA -0.21872146 -0.24536341 -0.1021407 -0.337272727
#> Patient_14 -0.36143106 -0.38923611 -0.41310795 -0.3630653 -0.331949250
#> Patient_15 -0.04147105          NA  0.02671756 -0.0679878           NA
#> Patient_16  0.40495495  0.41744186  0.48087331         NA           NA
#> Patient_17  0.72426598  0.74329004  0.75884265  0.8075163  0.727763100
#> Patient_18 -0.81907001          NA -0.80852594 -0.8341941           NA
#> Patient_19          NA  0.18406593  0.16672525  0.3896755  0.078541290
#> Patient_20          NA          NA  0.57878788         NA  0.549417064
#> Patient_21 -0.63349032 -0.67111111 -0.61812016 -0.5012821 -0.659071118
#> Patient_22  0.25073260  0.18561230  0.31785714  0.3691919  0.237529138
#> Patient_23 -0.59876847          NA          NA -0.4314815 -0.564191419
#> Patient_24 -0.44522059 -0.37316999          NA -0.2858887 -0.483764633
#> Patient_25  0.93294574          NA  0.93282636         NA  0.941176471
#> Patient_26          NA  0.66914894  0.58947368  0.7778336  0.611991870
#> Patient_27          NA -0.85490798 -0.80864098         NA -0.823744292
#> Patient_28  0.33884514  0.30014749  0.33200213         NA           NA
#> Patient_29          NA -0.76653543 -0.75425590 -0.7073681           NA
#> Patient_30 -0.44199561 -0.39658120          NA         NA -0.449301920
#>                     [,6]        [,7]        [,8]        [,9]       [,10]
#> Patient_1  -0.1088614116 -0.18307653 -0.19583333 -0.26076280          NA
#> Patient_2             NA -0.45793651 -0.45682261 -0.46971831 -0.46025119
#> Patient_3   0.8367579909  0.82009804  0.81376812  0.79751244          NA
#> Patient_4  -0.3013937282 -0.48686869 -0.43317560 -0.51273533 -0.45149254
#> Patient_5   0.8371522867          NA  0.83187135          NA  0.81907119
#> Patient_6             NA -0.63201320          NA -0.66076389          NA
#> Patient_7             NA  0.05120925  0.17163743  0.10388640          NA
#> Patient_8   0.3913019989          NA          NA  0.15998117  0.25855457
#> Patient_9   0.2992202729          NA  0.06007463  0.02519380  0.19171800
#> Patient_10            NA          NA -0.52750583 -0.55505395 -0.42454545
#> Patient_11 -0.5535676251          NA -0.56114473          NA -0.65928962
#> Patient_12            NA  0.18298368          NA  0.14880311  0.20064441
#> Patient_13 -0.1507299270 -0.23584906 -0.20530303 -0.19659091 -0.14652015
#> Patient_14 -0.2739669421 -0.42493113 -0.39425770 -0.42398374 -0.37743056
#> Patient_15  0.0007307874 -0.06978679          NA -0.07272727  0.08133748
#> Patient_16  0.5879915306  0.43986355  0.43681319  0.42506266  0.47172285
#> Patient_17            NA  0.76159420  0.73656126  0.74241439  0.76543450
#> Patient_18 -0.7758169935          NA -0.77150286 -0.83411215 -0.76386293
#> Patient_19  0.3472789116  0.10357363          NA          NA          NA
#> Patient_20            NA  0.53388704  0.63770227  0.56504065          NA
#> Patient_21 -0.6134005764          NA -0.65265487          NA          NA
#> Patient_22  0.3465478212  0.25740741  0.22923497  0.23606061  0.35286411
#> Patient_23 -0.4137681159 -0.62978003          NA          NA -0.50318841
#> Patient_24 -0.1500000000 -0.46604010 -0.39376812          NA -0.29128788
#> Patient_25            NA  0.92929293          NA  0.94855709  0.94358974
#> Patient_26  0.7173076923          NA  0.66604850  0.49241774          NA
#> Patient_27 -0.7904761905 -0.80227743 -0.84340010 -0.83020833 -0.78419913
#> Patient_28  0.4024590164  0.32595005  0.37521368          NA  0.41694537
#> Patient_29 -0.7172413793 -0.73469945          NA -0.74207921 -0.73389640
#> Patient_30 -0.2664921466 -0.44572271 -0.39839909          NA -0.42876053
#>                   [,11]        [,12]       [,13]       [,14]       [,15]
#> Patient_1  -0.414740369 -0.271903323 -0.28639241 -0.26482633 -0.20204978
#> Patient_2            NA -0.482822086          NA -0.44188034 -0.40960452
#> Patient_3   0.799539171  0.833229814  0.81502276          NA  0.81295488
#> Patient_4  -0.531369151 -0.551091107 -0.62267760 -0.48937762 -0.44636265
#> Patient_5   0.744117647  0.819523100  0.80427099  0.79418029          NA
#> Patient_6            NA -0.660176991 -0.69581240          NA -0.57794028
#> Patient_7  -0.003770739 -0.008199122  0.02878788  0.05481928          NA
#> Patient_8   0.026611418  0.127746741  0.05425532  0.18525641  0.20589354
#> Patient_9  -0.107742782           NA -0.04467456  0.03991935          NA
#> Patient_10 -0.569491525 -0.539583333 -0.54042357 -0.50594578 -0.46156053
#> Patient_11 -0.705784469 -0.705284553 -0.70584219 -0.70241851 -0.63267717
#> Patient_12           NA  0.165450122  0.03245614          NA  0.22479241
#> Patient_13 -0.372727273           NA -0.34231884 -0.38633399 -0.18374656
#> Patient_14           NA -0.426965003 -0.44364548          NA -0.37350427
#> Patient_15           NA -0.094736842          NA -0.13502825 -0.02915377
#> Patient_16  0.377468522  0.337646334  0.39147005  0.43955901  0.45651629
#> Patient_17  0.713745020  0.754889179  0.73311258  0.74122807  0.75512821
#> Patient_18 -0.841783671           NA          NA -0.80569620 -0.78320453
#> Patient_19 -0.086486486           NA -0.01140351  0.03825758          NA
#> Patient_20  0.564244521  0.553694581  0.55055741  0.53684211  0.56190476
#> Patient_21 -0.730000000 -0.605378678          NA -0.67384155 -0.56571219
#> Patient_22  0.069767442           NA          NA          NA          NA
#> Patient_23           NA -0.647038234 -0.62659799 -0.65064935 -0.49137645
#> Patient_24 -0.621907374 -0.511421320          NA -0.51447288          NA
#> Patient_25  0.928121059  0.959390863  0.93997446  0.94845361  0.96060991
#> Patient_26  0.493191057  0.564011799          NA  0.57176560  0.67472306
#> Patient_27           NA -0.785345717 -0.83197382          NA -0.77554842
#> Patient_28  0.118161683           NA  0.33008130  0.31505017          NA
#> Patient_29 -0.796564885           NA          NA          NA          NA
#> Patient_30           NA           NA -0.49120292          NA -0.40687646
#>                  [,16]       [,17]       [,18]       [,19]       [,20]
#> Patient_1  -0.25099623 -0.25562771          NA -0.06799687 -0.12960596
#> Patient_2           NA -0.47725989          NA -0.39863732 -0.36993464
#> Patient_3   0.82206235          NA  0.81328862  0.81536797  0.83971631
#> Patient_4  -0.42748344 -0.45984456 -0.52538860 -0.36003236          NA
#> Patient_5   0.79677950  0.76996124  0.77338129          NA          NA
#> Patient_6  -0.58801992          NA -0.66768707 -0.54572650 -0.55431300
#> Patient_7           NA          NA          NA  0.18982301          NA
#> Patient_8   0.23180212  0.12369250  0.09175627  0.32417840  0.29996726
#> Patient_9   0.09465409  0.02535934          NA  0.19682540          NA
#> Patient_10 -0.52145422 -0.52180685          NA          NA -0.33496241
#> Patient_11          NA          NA          NA -0.65137131 -0.63989024
#> Patient_12          NA          NA  0.06684588  0.20711893  0.31834247
#> Patient_13 -0.23202247 -0.30924504 -0.37238095          NA -0.06737589
#> Patient_14 -0.41596120 -0.44383302 -0.42677400 -0.29674797 -0.31596120
#> Patient_15 -0.21071429 -0.19831224 -0.12016985 -0.05864106  0.01428571
#> Patient_16  0.44429098  0.42432950  0.42430690  0.46735242  0.55348259
#> Patient_17  0.72817869  0.71634072  0.73726003  0.79551451          NA
#> Patient_18 -0.80937500 -0.83934681 -0.81728111 -0.80580848 -0.75275482
#> Patient_19  0.12941176  0.03410853 -0.06775362          NA  0.29841572
#> Patient_20  0.63213166  0.53887070  0.51885522          NA          NA
#> Patient_21 -0.66286727          NA -0.64275362 -0.62223065 -0.53157895
#> Patient_22  0.18436426  0.26784083  0.12812500  0.27128713          NA
#> Patient_23          NA          NA -0.67899250 -0.43113208 -0.57020202
#> Patient_24 -0.35525114 -0.50287356 -0.50610913          NA -0.26106613
#> Patient_25          NA  0.93877551  0.93519695          NA  0.95090439
#> Patient_26  0.62843557  0.53674541          NA  0.72222222  0.69160305
#> Patient_27          NA          NA          NA -0.81139241          NA
#> Patient_28          NA  0.21767338  0.31338798  0.41256831  0.43971880
#> Patient_29 -0.74802131 -0.77845011 -0.79488189          NA -0.73964346
#> Patient_30 -0.28062827 -0.47065217 -0.46074600 -0.30635739 -0.36116098
#> 
#> 
#> $BOOTSTRAP_VAR
#> $BOOTSTRAP_VAR$`2018_Tiriac_PDAC_PDO_classical_signature`
#>  Patient_1  Patient_2  Patient_3  Patient_4  Patient_5  Patient_6  Patient_7 
#> 0.07943830 0.05790320 0.02555551 0.08250896 0.03454246 0.04655894 0.11626292 
#>  Patient_8  Patient_9 Patient_10 Patient_11 Patient_12 Patient_13 Patient_14 
#> 0.12474454 0.07865670 0.06135358 0.04292565 0.07092631 0.09944061 0.05597327 
#> Patient_15 Patient_16 Patient_17 Patient_18 Patient_19 Patient_20 Patient_21 
#> 0.07291955 0.05787288 0.02773772 0.04174579 0.12749023 0.02773086 0.06392988 
#> Patient_22 Patient_23 Patient_24 Patient_25 Patient_26 Patient_27 Patient_28 
#> 0.06372530 0.06934161 0.10828696 0.01785433 0.07336527 0.02145951 0.06812883 
#> Patient_29 Patient_30 
#> 0.03306222 0.06194896 
#> 
#> $BOOTSTRAP_VAR$`2018_Tiriac_PDAC_PDO_basal-like_signature`
#>  Patient_1  Patient_2  Patient_3  Patient_4  Patient_5  Patient_6  Patient_7 
#> 0.08278513 0.03947803 0.01195197 0.08382162 0.02769437 0.05265411 0.09504494 
#>  Patient_8  Patient_9 Patient_10 Patient_11 Patient_12 Patient_13 Patient_14 
#> 0.10097610 0.11787536 0.05896700 0.05130093 0.09737175 0.09797071 0.05101063 
#> Patient_15 Patient_16 Patient_17 Patient_18 Patient_19 Patient_20 Patient_21 
#> 0.08028556 0.05815414 0.02495715 0.02862807 0.15100776 0.03547747 0.05818810 
#> Patient_22 Patient_23 Patient_24 Patient_25 Patient_26 Patient_27 Patient_28 
#> 0.08279678 0.09272931 0.12469251 0.01036520 0.08558314 0.02536408 0.08442939 
#> Patient_29 Patient_30 
#> 0.02832613 0.07153430 
#> 
#>